Update pyodide to 0.25.0
And tidy up loading process:

1. switch Python metadata from a user js module to an internal json module
2. switch lockfile.js to a json file and add capability for builtin json modules
