--- src/pyodide/internal/pyodide-lock.js
@@ -1,4311 +1,0 @@
-export const lockFile = {
-  "info":{
-    "arch":"wasm32",
-    "platform":"emscripten_3_1_45",
-    "python":"3.11.3",
-    "version":"0.25.0a2"
- },
- "packages":{
-    "aiohttp":{
-       "depends":[
-          "aiosignal",
-          "async-timeout",
-          "attrs",
-          "charset-normalizer",
-          "frozenlist",
-          "multidict",
-          "yarl"
-       ],
-       "file_name":"aiohttp-3.8.6-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "aiohttp"
-       ],
-       "install_dir":"site",
-       "name":"aiohttp",
-       "package_type":"package",
-       "sha256":"854fd5c0610779fda4b15af0c7ad72e2ebc196901cb460b8743f4ab6f58437cb",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"3.8.6"
-    },
-    "aiohttp-tests":{
-       "depends":[
-          "aiohttp"
-       ],
-       "file_name":"aiohttp-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"aiohttp-tests",
-       "package_type":"package",
-       "sha256":"1ab2c053eacc8b66f11aa47f6b86d3a138c05d4483d4e99cd0f603c190e176b0",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"3.8.6"
-    },
-    "aiosignal":{
-       "depends":[
-          "frozenlist"
-       ],
-       "file_name":"aiosignal-1.3.1-py3-none-any.whl",
-       "imports":[
-          "aiosignal"
-       ],
-       "install_dir":"site",
-       "name":"aiosignal",
-       "package_type":"package",
-       "sha256":"fe100169fa086824836f032f332e08e277a2e421c9a0a3efb9e82360e8fa6bd2",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.3.1"
-    },
-    "asciitree":{
-       "depends":[
-
-       ],
-       "file_name":"asciitree-0.3.3-py3-none-any.whl",
-       "imports":[
-          "asciitree"
-       ],
-       "install_dir":"site",
-       "name":"asciitree",
-       "package_type":"package",
-       "sha256":"fb26507dc81b3f232443cc9416fc80480359028595170bde91eea7696df0be06",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"0.3.3"
-    },
-    "astropy":{
-       "depends":[
-          "distutils",
-          "packaging",
-          "numpy",
-          "pyerfa",
-          "pyyaml"
-       ],
-       "file_name":"astropy-5.3.2-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "astropy"
-       ],
-       "install_dir":"site",
-       "name":"astropy",
-       "package_type":"package",
-       "sha256":"9834027485c4896ea9941576239d375a76c0a3e65c8ecb7b0a769c7a92725f25",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"5.3.2"
-    },
-    "async-timeout":{
-       "depends":[
-
-       ],
-       "file_name":"async_timeout-4.0.3-py3-none-any.whl",
-       "imports":[
-          "async_timeout"
-       ],
-       "install_dir":"site",
-       "name":"async-timeout",
-       "package_type":"package",
-       "sha256":"8d28f6341ed56de528a2e3692b4320392e99d096306341de7af1e851f0b944a1",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"4.0.3"
-    },
-    "atomicwrites":{
-       "depends":[
-
-       ],
-       "file_name":"atomicwrites-1.4.1-py2.py3-none-any.whl",
-       "imports":[
-          "atomicwrites"
-       ],
-       "install_dir":"site",
-       "name":"atomicwrites",
-       "package_type":"package",
-       "sha256":"ae2b4899ce5e50eea00675059d2f754de8163a2c1864f540b83321f3e3828cfc",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.4.1"
-    },
-    "attrs":{
-       "depends":[
-          "six"
-       ],
-       "file_name":"attrs-23.1.0-py3-none-any.whl",
-       "imports":[
-          "attr",
-          "attrs"
-       ],
-       "install_dir":"site",
-       "name":"attrs",
-       "package_type":"package",
-       "sha256":"7386045958ed3ac79bf113a9a91b7eb15b5bb050d3b9d185b150d97c9af42462",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"23.1.0"
-    },
-    "autograd":{
-       "depends":[
-          "numpy",
-          "future"
-       ],
-       "file_name":"autograd-1.6.2-py3-none-any.whl",
-       "imports":[
-          "autograd"
-       ],
-       "install_dir":"site",
-       "name":"autograd",
-       "package_type":"package",
-       "sha256":"8e16efbda828fc35e026b541c1b073ba97b47fc15cdaeea8f6066222a47b3f6e",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"1.6.2"
-    },
-    "autograd-tests":{
-       "depends":[
-          "autograd"
-       ],
-       "file_name":"autograd-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"autograd-tests",
-       "package_type":"package",
-       "sha256":"853cc16666c6a14b9f2b49e62a4386b4b2b4d8491b965662f6ff5ebead5c77d4",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.6.2"
-    },
-    "awkward-cpp":{
-       "depends":[
-          "numpy"
-       ],
-       "file_name":"awkward_cpp-26-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "awkward_cpp"
-       ],
-       "install_dir":"site",
-       "name":"awkward-cpp",
-       "package_type":"package",
-       "sha256":"a51896d9e2f5481d39f359bd97f34d12dac560c5b44fac8e9f9e3d52ecb3845b",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"26"
-    },
-    "b2d":{
-       "depends":[
-          "numpy",
-          "pydantic",
-          "setuptools"
-       ],
-       "file_name":"b2d-0.7.4-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "b2d"
-       ],
-       "install_dir":"site",
-       "name":"b2d",
-       "package_type":"package",
-       "sha256":"2c3956f7a7c04b8d61002612d5cf9ad1d5f2b5a04630700eb475bc94a3652d74",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"0.7.4"
-    },
-    "bcrypt":{
-       "depends":[
-
-       ],
-       "file_name":"bcrypt-4.0.1-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "bcrypt"
-       ],
-       "install_dir":"site",
-       "name":"bcrypt",
-       "package_type":"package",
-       "sha256":"d55db581585f0557c2fe3622cb1bf09503fb31edabd23e2e8e2caa79557cfbba",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"4.0.1"
-    },
-    "beautifulsoup4":{
-       "depends":[
-          "soupsieve"
-       ],
-       "file_name":"beautifulsoup4-4.12.2-py3-none-any.whl",
-       "imports":[
-          "bs4"
-       ],
-       "install_dir":"site",
-       "name":"beautifulsoup4",
-       "package_type":"package",
-       "sha256":"4a253cd25842a3d175570bbeb1d250e7769862340bcc4efeda683aece76e6dfd",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"4.12.2"
-    },
-    "beautifulsoup4-tests":{
-       "depends":[
-          "beautifulsoup4"
-       ],
-       "file_name":"beautifulsoup4-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"beautifulsoup4-tests",
-       "package_type":"package",
-       "sha256":"93e4dd3528cd496f9aaf56c1f09775beabbda187715eb9c3c13baae8077bf43c",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"4.12.2"
-    },
-    "biopython":{
-       "depends":[
-          "numpy"
-       ],
-       "file_name":"biopython-1.81-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "Bio",
-          "BioSQL"
-       ],
-       "install_dir":"site",
-       "name":"biopython",
-       "package_type":"package",
-       "sha256":"1879610535d53d90d38da128ae277e893da683568444caa11e360a9a1144774d",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.81"
-    },
-    "bitarray":{
-       "depends":[
-
-       ],
-       "file_name":"bitarray-2.8.1-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "bitarray"
-       ],
-       "install_dir":"site",
-       "name":"bitarray",
-       "package_type":"package",
-       "sha256":"80689f5702a9a92061b9aa5a42e0d3e2966ebf68f6ba01d9ba35a18b92458b89",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"2.8.1"
-    },
-    "bitarray-tests":{
-       "depends":[
-          "bitarray"
-       ],
-       "file_name":"bitarray-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"bitarray-tests",
-       "package_type":"package",
-       "sha256":"400ece37cd1a68543a9432edc4db7b520798087a63d472204a489ba734b1f939",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"2.8.1"
-    },
-    "bitstring":{
-       "depends":[
-          "bitarray"
-       ],
-       "file_name":"bitstring-4.1.1-py3-none-any.whl",
-       "imports":[
-          "bitstring"
-       ],
-       "install_dir":"site",
-       "name":"bitstring",
-       "package_type":"package",
-       "sha256":"6d1da3b3504386eb9463e75077e0b219c77037ef3d69885d3369dc01bb45b1cf",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"4.1.1"
-    },
-    "bleach":{
-       "depends":[
-          "webencodings",
-          "packaging",
-          "six"
-       ],
-       "file_name":"bleach-6.0.0-py3-none-any.whl",
-       "imports":[
-          "bleach"
-       ],
-       "install_dir":"site",
-       "name":"bleach",
-       "package_type":"package",
-       "sha256":"b1f4be00e20c6997de2540b734d813a69a03577ed23d553a717e7e052a583e78",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"6.0.0"
-    },
-    "bokeh":{
-       "depends":[
-          "distutils",
-          "numpy",
-          "jinja2",
-          "pandas",
-          "pillow",
-          "python-dateutil",
-          "six",
-          "typing-extensions",
-          "pyyaml",
-          "xyzservices"
-       ],
-       "file_name":"bokeh-3.2.2-py3-none-any.whl",
-       "imports":[
-          "bokeh"
-       ],
-       "install_dir":"site",
-       "name":"bokeh",
-       "package_type":"package",
-       "sha256":"4ad97fab8b0ae9d2ca32d9cd3bc434776948bc0692cf5a4a85bc23a76d6d5da0",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"3.2.2"
-    },
-    "boost-histogram":{
-       "depends":[
-          "numpy"
-       ],
-       "file_name":"boost_histogram-1.3.2-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "boost_histogram"
-       ],
-       "install_dir":"site",
-       "name":"boost-histogram",
-       "package_type":"package",
-       "sha256":"75d83324cbdde87376422f4c7a49a368c5db2cb4baa1c456d29eb423360a11c9",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.3.2"
-    },
-    "brotli":{
-       "depends":[
-
-       ],
-       "file_name":"Brotli-1.0.9-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "brotli"
-       ],
-       "install_dir":"site",
-       "name":"brotli",
-       "package_type":"package",
-       "sha256":"e663cfc5e7e832f74bbb79fa5e04cde619b7241c1d74e95b3d3c806a61630454",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.0.9"
-    },
-    "cachetools":{
-       "depends":[
-
-       ],
-       "file_name":"cachetools-5.3.1-py3-none-any.whl",
-       "imports":[
-          "cachetools"
-       ],
-       "install_dir":"site",
-       "name":"cachetools",
-       "package_type":"package",
-       "sha256":"13a4549dc4e04bbd93d5cdf25b907cbce7bd3f3053f6e4cee13bac50bf4fbbc3",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"5.3.1"
-    },
-    "cartopy":{
-       "depends":[
-          "shapely",
-          "pyshp",
-          "pyproj",
-          "geos",
-          "matplotlib",
-          "scipy"
-       ],
-       "file_name":"Cartopy-0.21.1-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "cartopy"
-       ],
-       "install_dir":"site",
-       "name":"Cartopy",
-       "package_type":"package",
-       "sha256":"e9077a93f1cd0bfd481de4dbc4cb8c6a2bb77e9b688dd4915ce53a4642d86fff",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"0.21.1"
-    },
-    "cartopy-tests":{
-       "depends":[
-          "cartopy"
-       ],
-       "file_name":"Cartopy-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"Cartopy-tests",
-       "package_type":"package",
-       "sha256":"bb5b757afe1534494064395865223048c361c9ff24c6bd933762240206d384b7",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"0.21.1"
-    },
-    "cbor-diag":{
-       "depends":[
-
-       ],
-       "file_name":"cbor_diag-1.0.1-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "cbor_diag"
-       ],
-       "install_dir":"site",
-       "name":"cbor-diag",
-       "package_type":"package",
-       "sha256":"85fe0547cc996e251311531b0c92e6667b51c179a0f229ddb94fa8e879980ee8",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.0.1"
-    },
-    "certifi":{
-       "depends":[
-
-       ],
-       "file_name":"certifi-2023.7.22-py3-none-any.whl",
-       "imports":[
-          "certifi"
-       ],
-       "install_dir":"site",
-       "name":"certifi",
-       "package_type":"package",
-       "sha256":"547743a4a25572b995e27837ff82d5434d86f8d808d9b98515b90d0eb5dbea8d",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"2023.7.22"
-    },
-    "cffi":{
-       "depends":[
-          "pycparser"
-       ],
-       "file_name":"cffi-1.15.1-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "cffi"
-       ],
-       "install_dir":"site",
-       "name":"cffi",
-       "package_type":"package",
-       "sha256":"86ea1797df618dc7505c9416742c3d3e19aa3dbec713206a3801c74527e8389f",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.15.1"
-    },
-    "cffi_example":{
-       "depends":[
-          "cffi"
-       ],
-       "file_name":"cffi_example-0.1-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "cffi_example"
-       ],
-       "install_dir":"site",
-       "name":"cffi_example",
-       "package_type":"package",
-       "sha256":"d89bc060961edf8c08a508f236fd0b3617b870f50e31dd593da5eebbda2d3a3d",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"0.1"
-    },
-    "cftime":{
-       "depends":[
-          "numpy"
-       ],
-       "file_name":"cftime-1.6.2-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "cftime"
-       ],
-       "install_dir":"site",
-       "name":"cftime",
-       "package_type":"package",
-       "sha256":"66b2110d74571bcd1f06b85bbc6c3c950701ac9a518c61238a2f5f22427337b8",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.6.2"
-    },
-    "charset-normalizer":{
-       "depends":[
-
-       ],
-       "file_name":"charset_normalizer-3.3.2-py3-none-any.whl",
-       "imports":[
-          "charset_normalizer"
-       ],
-       "install_dir":"site",
-       "name":"charset-normalizer",
-       "package_type":"package",
-       "sha256":"998f8057f9cb9594e41f81fb8775a48e79bc672dbdc8f97756d514e35da336fa",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"3.3.2"
-    },
-    "click":{
-       "depends":[
-
-       ],
-       "file_name":"click-8.1.7-py3-none-any.whl",
-       "imports":[
-          "click"
-       ],
-       "install_dir":"site",
-       "name":"click",
-       "package_type":"package",
-       "sha256":"1f9b6161df3a49d19552bfa4a049cf10d4aa7f73fd3a225a7fef346cddcfdccf",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"8.1.7"
-    },
-    "cligj":{
-       "depends":[
-          "click"
-       ],
-       "file_name":"cligj-0.7.2-py3-none-any.whl",
-       "imports":[
-          "cligj"
-       ],
-       "install_dir":"site",
-       "name":"cligj",
-       "package_type":"package",
-       "sha256":"b6a6b729fa7b944c77834bf9c9d5d26e7797116fcadeafb43896e5054db86cdd",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"0.7.2"
-    },
-    "cloudpickle":{
-       "depends":[
-
-       ],
-       "file_name":"cloudpickle-2.2.1-py3-none-any.whl",
-       "imports":[
-          "cloudpickle"
-       ],
-       "install_dir":"site",
-       "name":"cloudpickle",
-       "package_type":"package",
-       "sha256":"d0be6900d8ece0f40491c8901991e699ead0def879b70ee8e9847539e3c3d4da",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"2.2.1"
-    },
-    "cmyt":{
-       "depends":[
-          "colorspacious",
-          "matplotlib",
-          "more-itertools",
-          "numpy"
-       ],
-       "file_name":"cmyt-1.3.1-py3-none-any.whl",
-       "imports":[
-          "cmyt"
-       ],
-       "install_dir":"site",
-       "name":"cmyt",
-       "package_type":"package",
-       "sha256":"1ce56c6d66e1b6b9ecc66741359962f72c2c26bb74a27870a440fc9863635fb1",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"1.3.1"
-    },
-    "cmyt-tests":{
-       "depends":[
-          "cmyt"
-       ],
-       "file_name":"cmyt-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"cmyt-tests",
-       "package_type":"package",
-       "sha256":"62594757d81a8ce6169a6183cff25a9b8f6e02fb7e31a93e5487539c02d5c4d8",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.3.1"
-    },
-    "colorspacious":{
-       "depends":[
-          "numpy"
-       ],
-       "file_name":"colorspacious-1.1.2-py2.py3-none-any.whl",
-       "imports":[
-          "colorspacious"
-       ],
-       "install_dir":"site",
-       "name":"colorspacious",
-       "package_type":"package",
-       "sha256":"4697909f872afd9256dc49c5f9dae5c7f192fea346885755eab261537e79bb7c",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.1.2"
-    },
-    "contourpy":{
-       "depends":[
-          "numpy"
-       ],
-       "file_name":"contourpy-1.2.0-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "contourpy"
-       ],
-       "install_dir":"site",
-       "name":"contourpy",
-       "package_type":"package",
-       "sha256":"70175343b93267850c1d3e9ef598136d24f11f2e390952c10059fcb72d5faecd",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.2.0"
-    },
-    "coolprop":{
-       "depends":[
-          "numpy",
-          "matplotlib"
-       ],
-       "file_name":"CoolProp-6.4.3-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "CoolProp"
-       ],
-       "install_dir":"site",
-       "name":"coolprop",
-       "package_type":"package",
-       "sha256":"cbae3ae545a718aa4d7a0ddeb8ec9b83df31413778e6ee3b344a447041886d87",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"6.4.3"
-    },
-    "coolprop-tests":{
-       "depends":[
-          "coolprop"
-       ],
-       "file_name":"coolprop-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"coolprop-tests",
-       "package_type":"package",
-       "sha256":"5dabce200039a84809325d9f42707a2802e5b5d7c985a88dd74908c05516f7ba",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"6.4.3"
-    },
-    "coverage":{
-       "depends":[
-          "sqlite3"
-       ],
-       "file_name":"coverage-7.3.0-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "coverage"
-       ],
-       "install_dir":"site",
-       "name":"coverage",
-       "package_type":"package",
-       "sha256":"1af63e51eaf475abbd6444294f56c86741627f3889835bb06a52511e222adb7d",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"7.3.0"
-    },
-    "cpp-exceptions-test":{
-       "depends":[
-
-       ],
-       "file_name":"cpp-exceptions-test-0.1.zip",
-       "imports":[
-
-       ],
-       "install_dir":"dynlib",
-       "name":"cpp-exceptions-test",
-       "package_type":"shared_library",
-       "sha256":"7a899580edb61ac8fbcafbe07fa9c3d39460a967f7cfbde7960dceba66e9db26",
-       "shared_library":true,
-       "unvendored_tests":false,
-       "version":"0.1"
-    },
-    "cramjam":{
-       "depends":[
-
-       ],
-       "file_name":"cramjam-2.6.2-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "cramjam"
-       ],
-       "install_dir":"site",
-       "name":"cramjam",
-       "package_type":"package",
-       "sha256":"2a6bfdff4bc9730c36c73d6b97db9883bdeb25f75fa1bbd230d4080d3e02a5a3",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"2.6.2"
-    },
-    "cryptography":{
-       "depends":[
-          "openssl",
-          "six",
-          "cffi"
-       ],
-       "file_name":"cryptography-39.0.2-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "cryptography"
-       ],
-       "install_dir":"site",
-       "name":"cryptography",
-       "package_type":"package",
-       "sha256":"392265913d50b703c2604670adbb1a4775109bb1586cb6401adb1d8cba41c6df",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"39.0.2"
-    },
-    "cssselect":{
-       "depends":[
-
-       ],
-       "file_name":"cssselect-1.2.0-py2.py3-none-any.whl",
-       "imports":[
-          "cssselect"
-       ],
-       "install_dir":"site",
-       "name":"cssselect",
-       "package_type":"package",
-       "sha256":"8dfde5cc1cc9049ca1c6e47014931aa6fb75478b795e98c6d1ef3a01ecfe39a4",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.2.0"
-    },
-    "cycler":{
-       "depends":[
-          "six"
-       ],
-       "file_name":"cycler-0.11.0-py3-none-any.whl",
-       "imports":[
-          "cycler"
-       ],
-       "install_dir":"site",
-       "name":"cycler",
-       "package_type":"package",
-       "sha256":"8320ab5ce7427075ef94e054621a57dad1781373d1066727efb2730c3e8c312f",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"0.11.0"
-    },
-    "cytoolz":{
-       "depends":[
-          "nose",
-          "toolz"
-       ],
-       "file_name":"cytoolz-0.12.2-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "cytoolz"
-       ],
-       "install_dir":"site",
-       "name":"cytoolz",
-       "package_type":"package",
-       "sha256":"f0bc5ac082027428c56876b5bef1ea32a488771f114ebc955e415fc7585dbc2d",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"0.12.2"
-    },
-    "cytoolz-tests":{
-       "depends":[
-          "cytoolz"
-       ],
-       "file_name":"cytoolz-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"cytoolz-tests",
-       "package_type":"package",
-       "sha256":"87a54911c5a24cb81acfb4de896cbd1b10c90c9ad04201813a35d6cce90d3c27",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"0.12.2"
-    },
-    "decorator":{
-       "depends":[
-
-       ],
-       "file_name":"decorator-5.1.1-py3-none-any.whl",
-       "imports":[
-          "decorator"
-       ],
-       "install_dir":"site",
-       "name":"decorator",
-       "package_type":"package",
-       "sha256":"6539a0f84acca15fbb4d46a8a280b78666680ebd2e269d8870aa5bdb5837b71c",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"5.1.1"
-    },
-    "demes":{
-       "depends":[
-          "attrs",
-          "ruamel.yaml"
-       ],
-       "file_name":"demes-0.2.3-py3-none-any.whl",
-       "imports":[
-          "demes"
-       ],
-       "install_dir":"site",
-       "name":"demes",
-       "package_type":"package",
-       "sha256":"9000f7ab28d391b3eb1b7ad8600891384a3efa4c9dc26e63f63f41045136f86a",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"0.2.3"
-    },
-    "deprecation":{
-       "depends":[
-          "packaging"
-       ],
-       "file_name":"deprecation-2.1.0-py2.py3-none-any.whl",
-       "imports":[
-          "deprecation"
-       ],
-       "install_dir":"site",
-       "name":"deprecation",
-       "package_type":"package",
-       "sha256":"f89fc415bb71b2209fd4e795c65986531c05d72d2e64d510f4f217088d474a3d",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"2.1.0"
-    },
-    "distlib":{
-       "depends":[
-
-       ],
-       "file_name":"distlib-0.3.7-py2.py3-none-any.whl",
-       "imports":[
-          "distlib"
-       ],
-       "install_dir":"site",
-       "name":"distlib",
-       "package_type":"package",
-       "sha256":"a277cd34e1da10b8a3c0c05e05d5f24b9a377ce9bd678d4dbd542e0334c5e39e",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"0.3.7"
-    },
-    "distutils":{
-       "depends":[
-
-       ],
-       "file_name":"distutils-1.0.0.zip",
-       "imports":[
-          "distutils"
-       ],
-       "install_dir":"stdlib",
-       "name":"distutils",
-       "package_type":"cpython_module",
-       "sha256":"d8cadd70c2fc9e3440bfe80aedd8d1d1666f869c79a0ec7d72adcb5e2ee2f183",
-       "shared_library":true,
-       "unvendored_tests":false,
-       "version":"1.0.0"
-    },
-    "docutils":{
-       "depends":[
-
-       ],
-       "file_name":"docutils-0.20.1-py3-none-any.whl",
-       "imports":[
-          "docutils"
-       ],
-       "install_dir":"site",
-       "name":"docutils",
-       "package_type":"package",
-       "sha256":"0963c458fd65f26db587cb1bc85f33e57025be1301a5d7059e8725578d9af459",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"0.20.1"
-    },
-    "exceptiongroup":{
-       "depends":[
-
-       ],
-       "file_name":"exceptiongroup-1.1.3-py3-none-any.whl",
-       "imports":[
-          "exceptiongroup"
-       ],
-       "install_dir":"site",
-       "name":"exceptiongroup",
-       "package_type":"package",
-       "sha256":"1306e3d66968fca296b531345bbe78644ed1c871e036cf52d0be12e569b83567",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.1.3"
-    },
-    "fastparquet":{
-       "depends":[
-          "cramjam",
-          "numpy",
-          "pandas",
-          "fsspec",
-          "packaging"
-       ],
-       "file_name":"fastparquet-2023.7.0-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "fastparquet"
-       ],
-       "install_dir":"site",
-       "name":"fastparquet",
-       "package_type":"package",
-       "sha256":"d737adbcb097c9782f5945e8508f4f550a07cb3757eeef3670d4767ed9652aa2",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"2023.7.0"
-    },
-    "fiona":{
-       "depends":[
-          "attrs",
-          "certifi",
-          "setuptools",
-          "six",
-          "click",
-          "cligj"
-       ],
-       "file_name":"Fiona-1.8.22-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "fiona"
-       ],
-       "install_dir":"site",
-       "name":"fiona",
-       "package_type":"package",
-       "sha256":"b92ebf209d6887cfdf14b45bb0e7a410c97b1609a303f69627c37b0578a094a5",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.8.22"
-    },
-    "fonttools":{
-       "depends":[
-
-       ],
-       "file_name":"fonttools-4.42.1-py3-none-any.whl",
-       "imports":[
-          "fontTools"
-       ],
-       "install_dir":"site",
-       "name":"fonttools",
-       "package_type":"package",
-       "sha256":"5fa9c93bc21b06855024e322ae4da232475e66b8b541f1b0ff5c33822da47058",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"4.42.1"
-    },
-    "fpcast-test":{
-       "depends":[
-
-       ],
-       "file_name":"fpcast_test-0.1.1-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "fpcast_test"
-       ],
-       "install_dir":"site",
-       "name":"fpcast-test",
-       "package_type":"package",
-       "sha256":"1120280151760d20ec382f785b1d2b27e7995584a506ea94e991cbf8fea6ef0a",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"0.1.1"
-    },
-    "freesasa":{
-       "depends":[
-
-       ],
-       "file_name":"freesasa-2.2.0.post3-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "freesasa"
-       ],
-       "install_dir":"site",
-       "name":"freesasa",
-       "package_type":"package",
-       "sha256":"def9f1530244caae05483cc4a60ebf8623468feae2d0786c31a3a64d8b3779f7",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"2.2.0.post3"
-    },
-    "frozenlist":{
-       "depends":[
-
-       ],
-       "file_name":"frozenlist-1.4.0-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "frozenlist"
-       ],
-       "install_dir":"site",
-       "name":"frozenlist",
-       "package_type":"package",
-       "sha256":"fac922df373487bcb5aea344b0d23d828a308c0460a2c0ae3792a7e2f828f377",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.4.0"
-    },
-    "fsspec":{
-       "depends":[
-
-       ],
-       "file_name":"fsspec-2023.6.0-py3-none-any.whl",
-       "imports":[
-          "fsspec"
-       ],
-       "install_dir":"site",
-       "name":"fsspec",
-       "package_type":"package",
-       "sha256":"becc121a3e198f4ed197c8cf6d471a5288b0c598c365462c043a7bad34142cc6",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"2023.6.0"
-    },
-    "fsspec-tests":{
-       "depends":[
-          "fsspec"
-       ],
-       "file_name":"fsspec-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"fsspec-tests",
-       "package_type":"package",
-       "sha256":"f63c7387ae802b3e0d8ac81d13653c3be05c72e611ecca15b9fab27a4178a44d",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"2023.6.0"
-    },
-    "future":{
-       "depends":[
-
-       ],
-       "file_name":"future-0.18.3-py3-none-any.whl",
-       "imports":[
-          "future"
-       ],
-       "install_dir":"site",
-       "name":"future",
-       "package_type":"package",
-       "sha256":"60189fdda509e850c9911d8b995a04e6286399be6d773885b800b5c10cd4b9c8",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"0.18.3"
-    },
-    "future-tests":{
-       "depends":[
-          "future"
-       ],
-       "file_name":"future-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"future-tests",
-       "package_type":"package",
-       "sha256":"6886cd32e934cdf748295bd69b30b3046d79eb5625fe2518020d386fccef178f",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"0.18.3"
-    },
-    "galpy":{
-       "depends":[
-          "numpy",
-          "scipy",
-          "matplotlib",
-          "astropy",
-          "future",
-          "setuptools"
-       ],
-       "file_name":"galpy-1.9.0-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "galpy"
-       ],
-       "install_dir":"site",
-       "name":"galpy",
-       "package_type":"package",
-       "sha256":"f75c44d345e16cc990b4f8f3c3380ad520d16f637229ea347ef869bf00fa7873",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.9.0"
-    },
-    "gdal":{
-       "depends":[
-          "geos"
-       ],
-       "file_name":"gdal-3.5.1.zip",
-       "imports":[
-
-       ],
-       "install_dir":"dynlib",
-       "name":"gdal",
-       "package_type":"shared_library",
-       "sha256":"728042045a7d6dde8fbc1a32f7dbb878c98da5dfbd577dc6428ca6563e2f3434",
-       "shared_library":true,
-       "unvendored_tests":false,
-       "version":"3.5.1"
-    },
-    "gensim":{
-       "depends":[
-          "numpy",
-          "scipy",
-          "six",
-          "smart_open"
-       ],
-       "file_name":"gensim-4.3.1-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "gensim"
-       ],
-       "install_dir":"site",
-       "name":"gensim",
-       "package_type":"package",
-       "sha256":"48605460b39d3f26160950b6b7fe93ac276ff9c44d00433d1b12dff4a3d11934",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"4.3.1"
-    },
-    "gensim-tests":{
-       "depends":[
-          "gensim"
-       ],
-       "file_name":"gensim-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"gensim-tests",
-       "package_type":"package",
-       "sha256":"288b2c2e1523a5e0751d9dae5d1a4996a4597abe5ff85b40a70f116fadf446c3",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"4.3.1"
-    },
-    "geopandas":{
-       "depends":[
-          "shapely",
-          "fiona",
-          "pyproj",
-          "packaging",
-          "pandas"
-       ],
-       "file_name":"geopandas-0.13.2-py3-none-any.whl",
-       "imports":[
-          "geopandas"
-       ],
-       "install_dir":"site",
-       "name":"geopandas",
-       "package_type":"package",
-       "sha256":"e13121638d499df5ce5934b89a24ec1c43040753e8c22fc0a1c92faf2b742efa",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"0.13.2"
-    },
-    "geopandas-tests":{
-       "depends":[
-          "geopandas"
-       ],
-       "file_name":"geopandas-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"geopandas-tests",
-       "package_type":"package",
-       "sha256":"22877d50f258cb570ba9e86d3da4d692ccbd2d935a196331ac4f3432b81cde56",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"0.13.2"
-    },
-    "geos":{
-       "depends":[
-
-       ],
-       "file_name":"geos-3.10.3.zip",
-       "imports":[
-
-       ],
-       "install_dir":"dynlib",
-       "name":"geos",
-       "package_type":"shared_library",
-       "sha256":"55561ff53f8369347821fcd1cf30fceb01a12a988a6596dd674303c865068b2a",
-       "shared_library":true,
-       "unvendored_tests":false,
-       "version":"3.10.3"
-    },
-    "gmpy2":{
-       "depends":[
-
-       ],
-       "file_name":"gmpy2-2.1.5-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "gmpy2"
-       ],
-       "install_dir":"site",
-       "name":"gmpy2",
-       "package_type":"package",
-       "sha256":"043c9b5e4abe52518f502fa5a56fba0040aa5310ecea5dd6d9eca492c1c3ca8d",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"2.1.5"
-    },
-    "gsw":{
-       "depends":[
-          "numpy"
-       ],
-       "file_name":"gsw-3.6.16-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "gsw"
-       ],
-       "install_dir":"site",
-       "name":"gsw",
-       "package_type":"package",
-       "sha256":"2a90fd15837bf082a7910cab42e642decbe383940ebb526a7ce0773cccef1ec1",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"3.6.16"
-    },
-    "gsw-tests":{
-       "depends":[
-          "gsw"
-       ],
-       "file_name":"gsw-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"gsw-tests",
-       "package_type":"package",
-       "sha256":"2a7815563dd0fed00865656d3da39d890ef065f0c00c1144879582b88ec0da7d",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"3.6.16"
-    },
-    "h5py":{
-       "depends":[
-          "numpy",
-          "pkgconfig"
-       ],
-       "file_name":"h5py-3.7.0-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "h5py"
-       ],
-       "install_dir":"site",
-       "name":"h5py",
-       "package_type":"package",
-       "sha256":"11acf34caf10a0924d7fd9a4dbb9fc97d0069eb2a22ca3f78ba42f2c7b9545e0",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"3.7.0"
-    },
-    "h5py-tests":{
-       "depends":[
-          "h5py"
-       ],
-       "file_name":"h5py-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"h5py-tests",
-       "package_type":"package",
-       "sha256":"d7f4b1f756ffd5537c81ed18893bb80873d3ccaaba3c8b7a95f07c01a6861e3e",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"3.7.0"
-    },
-    "hashlib":{
-       "depends":[
-          "openssl"
-       ],
-       "file_name":"hashlib-1.0.0.zip",
-       "imports":[
-          "_hashlib"
-       ],
-       "install_dir":"stdlib",
-       "name":"hashlib",
-       "package_type":"cpython_module",
-       "sha256":"72129cb53e800f1c78e190d0b0b4a006ef63df8027a8f37437df8f2516e5a82e",
-       "shared_library":true,
-       "unvendored_tests":false,
-       "version":"1.0.0"
-    },
-    "html5lib":{
-       "depends":[
-          "webencodings",
-          "six"
-       ],
-       "file_name":"html5lib-1.1-py2.py3-none-any.whl",
-       "imports":[
-          "html5lib"
-       ],
-       "install_dir":"site",
-       "name":"html5lib",
-       "package_type":"package",
-       "sha256":"05d32aa50ad724ce1be6e427719d5624c69686984a1f945f90f012f5aa6599dc",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.1"
-    },
-    "idna":{
-       "depends":[
-
-       ],
-       "file_name":"idna-3.4-py3-none-any.whl",
-       "imports":[
-          "idna"
-       ],
-       "install_dir":"site",
-       "name":"idna",
-       "package_type":"package",
-       "sha256":"7d910323658c592ef83305f5189251f33fd58288349b4a40445c2d159bbdaefe",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"3.4"
-    },
-    "igraph":{
-       "depends":[
-          "texttable"
-       ],
-       "file_name":"igraph-0.10.6-cp39-abi3-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "igraph"
-       ],
-       "install_dir":"site",
-       "name":"igraph",
-       "package_type":"package",
-       "sha256":"9691cf80615b7aabad9b43cb4663daef3eee1f90c19e6a3604eec9ab898a5f93",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"0.10.6"
-    },
-    "imageio":{
-       "depends":[
-          "numpy",
-          "pillow"
-       ],
-       "file_name":"imageio-2.31.1-py3-none-any.whl",
-       "imports":[
-          "imageio"
-       ],
-       "install_dir":"site",
-       "name":"imageio",
-       "package_type":"package",
-       "sha256":"e7fd110720324c826bd7f66035eed45d392d9105346321e437aabb2e46a85bf9",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"2.31.1"
-    },
-    "iniconfig":{
-       "depends":[
-
-       ],
-       "file_name":"iniconfig-2.0.0-py3-none-any.whl",
-       "imports":[
-          "iniconfig"
-       ],
-       "install_dir":"site",
-       "name":"iniconfig",
-       "package_type":"package",
-       "sha256":"65626eae0e1503ed3b8bcf850513a6727d5b94f179fc94d7f98eaf0ccf17e45d",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"2.0.0"
-    },
-    "jedi":{
-       "depends":[
-          "parso"
-       ],
-       "file_name":"jedi-0.19.0-py2.py3-none-any.whl",
-       "imports":[
-          "jedi"
-       ],
-       "install_dir":"site",
-       "name":"jedi",
-       "package_type":"package",
-       "sha256":"4f4ae884153bed284a1842696592ae691650f011c3bed537dcefe35d5ba5e8f7",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"0.19.0"
-    },
-    "jedi-tests":{
-       "depends":[
-          "jedi"
-       ],
-       "file_name":"jedi-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"jedi-tests",
-       "package_type":"package",
-       "sha256":"1b4df53e6ebfd353a555719d808878816f8acc7777b45ccbe0c528872f090fb1",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"0.19.0"
-    },
-    "jinja2":{
-       "depends":[
-          "markupsafe"
-       ],
-       "file_name":"Jinja2-3.1.2-py3-none-any.whl",
-       "imports":[
-          "jinja2"
-       ],
-       "install_dir":"site",
-       "name":"Jinja2",
-       "package_type":"package",
-       "sha256":"f411eb0e076459c2abbe87cfcb63d5ef02ef4563957cbbdb8155ede3e4439b6a",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"3.1.2"
-    },
-    "joblib":{
-       "depends":[
-          "distutils"
-       ],
-       "file_name":"joblib-1.3.2-py3-none-any.whl",
-       "imports":[
-          "joblib"
-       ],
-       "install_dir":"site",
-       "name":"joblib",
-       "package_type":"package",
-       "sha256":"e4ffd9ea772c7941a4b8c3fdfb2c9e4bef6e4db38e5d8660b391b3a1fa43f2f0",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"1.3.2"
-    },
-    "joblib-tests":{
-       "depends":[
-          "joblib"
-       ],
-       "file_name":"joblib-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"joblib-tests",
-       "package_type":"package",
-       "sha256":"d033523cf68a2805393270419a2188d0f39d919326be8b9b2b2719da4e71d665",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.3.2"
-    },
-    "jsonschema":{
-       "depends":[
-          "attrs",
-          "pyrsistent"
-       ],
-       "file_name":"jsonschema-4.17.3-py3-none-any.whl",
-       "imports":[
-          "jsonschema"
-       ],
-       "install_dir":"site",
-       "name":"jsonschema",
-       "package_type":"package",
-       "sha256":"095d6d4265f6d3e39a0fd453f6c71e3899c4ad0bc5a137338adca397a79f8c3b",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"4.17.3"
-    },
-    "jsonschema-tests":{
-       "depends":[
-          "jsonschema"
-       ],
-       "file_name":"jsonschema-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"jsonschema-tests",
-       "package_type":"package",
-       "sha256":"def354e1d4a6728081459edb3f0a4ec295f669763aa654a15a1b853638cc3152",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"4.17.3"
-    },
-    "kiwisolver":{
-       "depends":[
-
-       ],
-       "file_name":"kiwisolver-1.4.4-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "kiwisolver"
-       ],
-       "install_dir":"site",
-       "name":"kiwisolver",
-       "package_type":"package",
-       "sha256":"c1d99133bcd0993c00002c87f457bf549659d8612b32cee960ff7e8b060675fd",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.4.4"
-    },
-    "lazy-object-proxy":{
-       "depends":[
-
-       ],
-       "file_name":"lazy_object_proxy-1.9.0-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "lazy_object_proxy"
-       ],
-       "install_dir":"site",
-       "name":"lazy-object-proxy",
-       "package_type":"package",
-       "sha256":"fa9e94277db2e5f12db3e0f2a656b78cbda2daaf7da298227df87d47bcfff53b",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.9.0"
-    },
-    "lazy_loader":{
-       "depends":[
-
-       ],
-       "file_name":"lazy_loader-0.3-py3-none-any.whl",
-       "imports":[
-          "lazy_loader"
-       ],
-       "install_dir":"site",
-       "name":"lazy_loader",
-       "package_type":"package",
-       "sha256":"eb7ac08870c2a6c7ba6bcaaa2b1fd3d9ddca6dbe6e4cab988c6eafe96166813b",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"0.3"
-    },
-    "lazy_loader-tests":{
-       "depends":[
-          "lazy_loader"
-       ],
-       "file_name":"lazy_loader-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"lazy_loader-tests",
-       "package_type":"package",
-       "sha256":"1abead0a2b9467ff27e2f052cf3906d9587ded6cb458e00d40dd378b7cecc57e",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"0.3"
-    },
-    "libhdf5":{
-       "depends":[
-
-       ],
-       "file_name":"libhdf5-1.12.1.zip",
-       "imports":[
-
-       ],
-       "install_dir":"dynlib",
-       "name":"libhdf5",
-       "package_type":"shared_library",
-       "sha256":"82758fc30e2293ba0c73e6bf4b56002fb7a5ac89aca3450a6adfd21b1068dc36",
-       "shared_library":true,
-       "unvendored_tests":false,
-       "version":"1.12.1"
-    },
-    "libheif":{
-       "depends":[
-
-       ],
-       "file_name":"libheif-1.12.0.zip",
-       "imports":[
-
-       ],
-       "install_dir":"dynlib",
-       "name":"libheif",
-       "package_type":"shared_library",
-       "sha256":"22ab7046a94d2c3e4e126fc36793f1b319d93ce346f15821f30e0b53f85a9014",
-       "shared_library":true,
-       "unvendored_tests":false,
-       "version":"1.12.0"
-    },
-    "libmagic":{
-       "depends":[
-
-       ],
-       "file_name":"libmagic-5.42.zip",
-       "imports":[
-
-       ],
-       "install_dir":"dynlib",
-       "name":"libmagic",
-       "package_type":"shared_library",
-       "sha256":"a129e9770a769487eb55eec74f4899b62577e651ce5bde3d12ae0d889d8d3625",
-       "shared_library":true,
-       "unvendored_tests":false,
-       "version":"5.42"
-    },
-    "libnetcdf":{
-       "depends":[
-
-       ],
-       "file_name":"libnetcdf-4.9.2.zip",
-       "imports":[
-
-       ],
-       "install_dir":"dynlib",
-       "name":"libnetcdf",
-       "package_type":"shared_library",
-       "sha256":"3c1d42331b6250428279ce4deb82ffbbba28d60a5ba4ef1b0ee9391964f61bfc",
-       "shared_library":true,
-       "unvendored_tests":false,
-       "version":"4.9.2"
-    },
-    "lightgbm":{
-       "depends":[
-          "numpy",
-          "scipy",
-          "scikit-learn"
-       ],
-       "file_name":"lightgbm-3.3.5-py3-none-any.whl",
-       "imports":[
-          "lightgbm"
-       ],
-       "install_dir":"site",
-       "name":"lightgbm",
-       "package_type":"package",
-       "sha256":"f8c26c69f44df137e2b5d12f89313c162d3eae68183116490fc0b1da9512b04b",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"3.3.5"
-    },
-    "logbook":{
-       "depends":[
-
-       ],
-       "file_name":"Logbook-1.6.0-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "logbook"
-       ],
-       "install_dir":"site",
-       "name":"logbook",
-       "package_type":"package",
-       "sha256":"8e9c68329edb82540bba96b19096cd7bc206fb69987d8aae3c04730f54ef6f31",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.6.0"
-    },
-    "lxml":{
-       "depends":[
-
-       ],
-       "file_name":"lxml-4.9.3-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "lxml"
-       ],
-       "install_dir":"site",
-       "name":"lxml",
-       "package_type":"package",
-       "sha256":"2bd90df5e4edc571bce34fc07b60134a280f7b11654cbccffebc57752be4b7f9",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"4.9.3"
-    },
-    "lzma":{
-       "depends":[
-
-       ],
-       "file_name":"lzma-1.0.0.zip",
-       "imports":[
-          "lzma",
-          "_lzma"
-       ],
-       "install_dir":"stdlib",
-       "name":"lzma",
-       "package_type":"cpython_module",
-       "sha256":"3f00fb8cb952334af632a235e3dac232bc50f90bbdffedb78fe2d0ab7e5c0878",
-       "shared_library":true,
-       "unvendored_tests":false,
-       "version":"1.0.0"
-    },
-    "markupsafe":{
-       "depends":[
-
-       ],
-       "file_name":"MarkupSafe-2.1.3-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "markupsafe"
-       ],
-       "install_dir":"site",
-       "name":"MarkupSafe",
-       "package_type":"package",
-       "sha256":"e1fe6162d64df9b7e675ba700189defc9808c2e9de3d5c626ff7f56ede0fb0d1",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"2.1.3"
-    },
-    "matplotlib":{
-       "depends":[
-          "cycler",
-          "fonttools",
-          "kiwisolver",
-          "numpy",
-          "packaging",
-          "pillow",
-          "pyparsing",
-          "python-dateutil",
-          "pytz",
-          "matplotlib-pyodide"
-       ],
-       "file_name":"matplotlib-3.5.2-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "pylab",
-          "mpl_toolkits",
-          "matplotlib"
-       ],
-       "install_dir":"site",
-       "name":"matplotlib",
-       "package_type":"package",
-       "sha256":"5af190a621d5169b704050a578b5c0c6a3b73116973d7980afd2fd332fb180fa",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"3.5.2"
-    },
-    "matplotlib-pyodide":{
-       "depends":[
-
-       ],
-       "file_name":"matplotlib_pyodide-0.2.0-py3-none-any.whl",
-       "imports":[
-          "matplotlib_pyodide"
-       ],
-       "install_dir":"site",
-       "name":"matplotlib-pyodide",
-       "package_type":"package",
-       "sha256":"aa773ea485c1b2cd559eea047a9b36dec324b1bc2e6a7dfe315dffc60aa2b158",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"0.2.0"
-    },
-    "matplotlib-tests":{
-       "depends":[
-          "matplotlib"
-       ],
-       "file_name":"matplotlib-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"matplotlib-tests",
-       "package_type":"package",
-       "sha256":"6a1a99de43c7fc533afc92de6298e957659135832741ceff71f06ced8d20647a",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"3.5.2"
-    },
-    "micropip":{
-       "depends":[
-          "packaging"
-       ],
-       "file_name":"micropip-0.5.0-py3-none-any.whl",
-       "imports":[
-          "micropip"
-       ],
-       "install_dir":"site",
-       "name":"micropip",
-       "package_type":"package",
-       "sha256":"15af913a2cedfcd84693ef14e67cfdfcd6691a68241159e26821dae755e5ba08",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"0.5.0"
-    },
-    "mne":{
-       "depends":[
-          "distutils",
-          "numpy",
-          "scipy",
-          "setuptools",
-          "decorator"
-       ],
-       "file_name":"mne-1.5.0-py3-none-any.whl",
-       "imports":[
-          "mne"
-       ],
-       "install_dir":"site",
-       "name":"mne",
-       "package_type":"package",
-       "sha256":"74087844d2789b69d37bf9ba99b32b633c73e268afd43303ce712c009576bc1c",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"1.5.0"
-    },
-    "mne-tests":{
-       "depends":[
-          "mne"
-       ],
-       "file_name":"mne-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"mne-tests",
-       "package_type":"package",
-       "sha256":"8f388b4ee518be5b95740e83ecae6165d7d8f534d1d9cee813034afab8a35386",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.5.0"
-    },
-    "more-itertools":{
-       "depends":[
-
-       ],
-       "file_name":"more_itertools-10.1.0-py3-none-any.whl",
-       "imports":[
-          "more_itertools"
-       ],
-       "install_dir":"site",
-       "name":"more-itertools",
-       "package_type":"package",
-       "sha256":"8e49503c0c6db99bdccd214b9a3e6ba4d868b580fd1bcb78550ae6696394b4e5",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"10.1.0"
-    },
-    "mpmath":{
-       "depends":[
-
-       ],
-       "file_name":"mpmath-1.3.0-py3-none-any.whl",
-       "imports":[
-          "mpmath"
-       ],
-       "install_dir":"site",
-       "name":"mpmath",
-       "package_type":"package",
-       "sha256":"8f6857216e07e4d098734ea7cf73f71bc622237567e5ef533ac6c7132f8e22c6",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"1.3.0"
-    },
-    "mpmath-tests":{
-       "depends":[
-          "mpmath"
-       ],
-       "file_name":"mpmath-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"mpmath-tests",
-       "package_type":"package",
-       "sha256":"cc2bd04d7937afd764a16ced4b6188954aa3f6c9b4f8472768ba26e2823e6985",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.3.0"
-    },
-    "msgpack":{
-       "depends":[
-
-       ],
-       "file_name":"msgpack-1.0.5-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "msgpack"
-       ],
-       "install_dir":"site",
-       "name":"msgpack",
-       "package_type":"package",
-       "sha256":"6abae98f247aaaddf601cfa26bbc6819ddf3523fc4813c7a934c31d1297850e5",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.0.5"
-    },
-    "msgspec":{
-       "depends":[
-
-       ],
-       "file_name":"msgspec-0.18.4-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "msgspec"
-       ],
-       "install_dir":"site",
-       "name":"msgspec",
-       "package_type":"package",
-       "sha256":"6d1a30d674c7ca35790fc457177e127b5c93111bb51da49200ec5f2c3e669654",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"0.18.4"
-    },
-    "msprime":{
-       "depends":[
-          "numpy",
-          "newick",
-          "tskit",
-          "demes"
-       ],
-       "file_name":"msprime-1.2.0-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "msprime"
-       ],
-       "install_dir":"site",
-       "name":"msprime",
-       "package_type":"package",
-       "sha256":"92db0eb46e39d823d6f4b65ff6c31d9f7fcd655252b909d0ded8f9bc945fd6e2",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.2.0"
-    },
-    "multidict":{
-       "depends":[
-
-       ],
-       "file_name":"multidict-6.0.4-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "multidict"
-       ],
-       "install_dir":"site",
-       "name":"multidict",
-       "package_type":"package",
-       "sha256":"ecc8490799421361ad9b49e93124a47461f0d4d382d33f853923cc1ec24648cc",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"6.0.4"
-    },
-    "munch":{
-       "depends":[
-          "setuptools",
-          "six"
-       ],
-       "file_name":"munch-4.0.0-py2.py3-none-any.whl",
-       "imports":[
-          "munch"
-       ],
-       "install_dir":"site",
-       "name":"munch",
-       "package_type":"package",
-       "sha256":"2f20ba18013117b4b20d17c8f220dfecf89ea70c76a5477ac6dbdc2bdad17755",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"4.0.0"
-    },
-    "mypy":{
-       "depends":[
-
-       ],
-       "file_name":"mypy-1.5.1-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "mypyc",
-          "mypy"
-       ],
-       "install_dir":"site",
-       "name":"mypy",
-       "package_type":"package",
-       "sha256":"bc6bf0cbbe3ad6ea833ac8e3d05af32a84d8177696613b82b0978b97aa002824",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"1.5.1"
-    },
-    "mypy-tests":{
-       "depends":[
-          "mypy"
-       ],
-       "file_name":"mypy-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"mypy-tests",
-       "package_type":"package",
-       "sha256":"d01e6aa71cbc96c43e0aa7fa49731cbd666e06de25fcbb78147fc3ffe84855bb",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.5.1"
-    },
-    "netcdf4":{
-       "depends":[
-          "numpy",
-          "packaging",
-          "h5py",
-          "cftime",
-          "certifi"
-       ],
-       "file_name":"netCDF4-1.6.4-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "netCDF4"
-       ],
-       "install_dir":"site",
-       "name":"netcdf4",
-       "package_type":"package",
-       "sha256":"d54ccc1195cdc839aa8551bf99e94405dc927ea4592138092c51daa7e717229e",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.6.4"
-    },
-    "networkx":{
-       "depends":[
-          "decorator",
-          "setuptools",
-          "matplotlib",
-          "numpy"
-       ],
-       "file_name":"networkx-3.1-py3-none-any.whl",
-       "imports":[
-          "networkx"
-       ],
-       "install_dir":"site",
-       "name":"networkx",
-       "package_type":"package",
-       "sha256":"7f2a3842978c9c03da05f5e43c900d480fa7103b614500ae6ea5a8572e979de7",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"3.1"
-    },
-    "networkx-tests":{
-       "depends":[
-          "networkx"
-       ],
-       "file_name":"networkx-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"networkx-tests",
-       "package_type":"package",
-       "sha256":"4097fa179befd9c5664605a709eadf81ace3d78ebf4b02dc9960cd578795f5be",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"3.1"
-    },
-    "newick":{
-       "depends":[
-
-       ],
-       "file_name":"newick-1.9.0-py2.py3-none-any.whl",
-       "imports":[
-          "newick"
-       ],
-       "install_dir":"site",
-       "name":"newick",
-       "package_type":"package",
-       "sha256":"2f6f5999473c6437d3983562f403f0c041fbbe7f48de31bac4aa58a809247167",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.9.0"
-    },
-    "nlopt":{
-       "depends":[
-          "numpy"
-       ],
-       "file_name":"nlopt-2.7.0-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "nlopt"
-       ],
-       "install_dir":"site",
-       "name":"nlopt",
-       "package_type":"package",
-       "sha256":"8972865b92aac5e066209f2cc8f897c98e9c751a457077d4fc735ce9b286073b",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"2.7.0"
-    },
-    "nltk":{
-       "depends":[
-          "regex",
-          "sqlite3"
-       ],
-       "file_name":"nltk-3.8.1-py3-none-any.whl",
-       "imports":[
-          "nltk"
-       ],
-       "install_dir":"site",
-       "name":"nltk",
-       "package_type":"package",
-       "sha256":"88fc2cdd22198de29f266a54d069d6cc8da60a28a2cce08d126d97e96d4fc45f",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"3.8.1"
-    },
-    "nltk-tests":{
-       "depends":[
-          "nltk"
-       ],
-       "file_name":"nltk-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"nltk-tests",
-       "package_type":"package",
-       "sha256":"edcec21cce941892161dd37ff0bc1e8ad20e5d4b831a9bbdfb325675739c4fbe",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"3.8.1"
-    },
-    "nose":{
-       "depends":[
-          "setuptools"
-       ],
-       "file_name":"nose-1.3.7-py3-none-any.whl",
-       "imports":[
-          "nose"
-       ],
-       "install_dir":"site",
-       "name":"nose",
-       "package_type":"package",
-       "sha256":"9dad6d8168192618108180d25f87fce54a21b2f6d81620de2e9dd3f71ce8bac6",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.3.7"
-    },
-    "numcodecs":{
-       "depends":[
-          "numpy",
-          "msgpack"
-       ],
-       "file_name":"numcodecs-0.11.0-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "numcodecs"
-       ],
-       "install_dir":"site",
-       "name":"numcodecs",
-       "package_type":"package",
-       "sha256":"8493bfb004c5cbb7d76e15a07800cfdb414bb0183f7b521a5b5144da17cd8cf2",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"0.11.0"
-    },
-    "numcodecs-tests":{
-       "depends":[
-          "numcodecs"
-       ],
-       "file_name":"numcodecs-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"numcodecs-tests",
-       "package_type":"package",
-       "sha256":"e4a3e1db742bdb5a094785c7fe64f2707282d8ab78a77263dd117d3cc9ab398c",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"0.11.0"
-    },
-    "numpy":{
-       "depends":[
-
-       ],
-       "file_name":"numpy-1.26.1-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "numpy"
-       ],
-       "install_dir":"site",
-       "name":"numpy",
-       "package_type":"package",
-       "sha256":"323c7770cb5a7b359e2228afc340cb1339447f00ec5fdc2ee48e3259099558aa",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"1.26.1"
-    },
-    "numpy-tests":{
-       "depends":[
-          "numpy"
-       ],
-       "file_name":"numpy-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"numpy-tests",
-       "package_type":"package",
-       "sha256":"77cd2ec423d0e70915c0571347f2af08b985596cca7fb6290bc64178f798d436",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.26.1"
-    },
-    "openblas":{
-       "depends":[
-
-       ],
-       "file_name":"openblas-0.3.23.zip",
-       "imports":[
-
-       ],
-       "install_dir":"dynlib",
-       "name":"openblas",
-       "package_type":"shared_library",
-       "sha256":"5bf713b60fe7e1059fde281a41453a7371a8af5ca12064c27077ee3265d52de2",
-       "shared_library":true,
-       "unvendored_tests":false,
-       "version":"0.3.23"
-    },
-    "opencv-python":{
-       "depends":[
-          "numpy"
-       ],
-       "file_name":"opencv_python-4.8.0.76-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "cv2"
-       ],
-       "install_dir":"site",
-       "name":"opencv-python",
-       "package_type":"package",
-       "sha256":"96cdd4d78f534082f7b3941a731c31b0aacf1965748a45f0d64963894e1d1c13",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"4.8.0.76"
-    },
-    "openssl":{
-       "depends":[
-
-       ],
-       "file_name":"openssl-1.1.1n.zip",
-       "imports":[
-
-       ],
-       "install_dir":"dynlib",
-       "name":"openssl",
-       "package_type":"shared_library",
-       "sha256":"c316da64fe4bad222a034a79425d2a63b8f64fe121acdecb684d74df06202180",
-       "shared_library":true,
-       "unvendored_tests":false,
-       "version":"1.1.1n"
-    },
-    "optlang":{
-       "depends":[
-          "sympy",
-          "six",
-          "swiglpk"
-       ],
-       "file_name":"optlang-1.7.0-py2.py3-none-any.whl",
-       "imports":[
-          "optlang"
-       ],
-       "install_dir":"site",
-       "name":"optlang",
-       "package_type":"package",
-       "sha256":"23c274ef947379bba89cfa45970a629d0bc58cf82bf9a63e732ed8e3720f0077",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"1.7.0"
-    },
-    "optlang-tests":{
-       "depends":[
-          "optlang"
-       ],
-       "file_name":"optlang-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"optlang-tests",
-       "package_type":"package",
-       "sha256":"d55a2a4b12cd2b808624f20a1a41ee830f0e1ab94d2d51c83b7967a68154778c",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.7.0"
-    },
-    "orjson":{
-       "depends":[
-
-       ],
-       "file_name":"orjson-3.9.2-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "orjson"
-       ],
-       "install_dir":"site",
-       "name":"orjson",
-       "package_type":"package",
-       "sha256":"a588f5a349ffe55785047912028e43327270d37f2672188e0d8c00fd4ba3cf1d",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"3.9.2"
-    },
-    "packaging":{
-       "depends":[
-
-       ],
-       "file_name":"packaging-23.1-py3-none-any.whl",
-       "imports":[
-          "packaging"
-       ],
-       "install_dir":"site",
-       "name":"packaging",
-       "package_type":"package",
-       "sha256":"1c52cd3001c09e0a9798d98a5669a39148b7f6732d54e01d10fadc8598cc2cce",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"23.1"
-    },
-    "pandas":{
-       "depends":[
-          "numpy",
-          "python-dateutil",
-          "pytz"
-       ],
-       "file_name":"pandas-1.5.3-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "pandas"
-       ],
-       "install_dir":"site",
-       "name":"pandas",
-       "package_type":"package",
-       "sha256":"6c7f383300d8c0224aff5ddb5e59a1662e00aee03e4f8ef9655032c0da7dd08b",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"1.5.3"
-    },
-    "pandas-tests":{
-       "depends":[
-          "pandas"
-       ],
-       "file_name":"pandas-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"pandas-tests",
-       "package_type":"package",
-       "sha256":"5f7fb8b0e9ca8856c093f39570014f3e1ae532f8d68de3835d9485beace7f8a7",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.5.3"
-    },
-    "parso":{
-       "depends":[
-
-       ],
-       "file_name":"parso-0.8.3-py2.py3-none-any.whl",
-       "imports":[
-          "parso"
-       ],
-       "install_dir":"site",
-       "name":"parso",
-       "package_type":"package",
-       "sha256":"534028219f51f0ae31543db279b1d4ebfec93fda1d413f6f33a8dad59369f475",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"0.8.3"
-    },
-    "patsy":{
-       "depends":[
-          "numpy",
-          "six"
-       ],
-       "file_name":"patsy-0.5.3-py2.py3-none-any.whl",
-       "imports":[
-          "patsy"
-       ],
-       "install_dir":"site",
-       "name":"patsy",
-       "package_type":"package",
-       "sha256":"68fce84ced6fdabc211e0bb15a65a8bc86fb1168e7d66dfcd7ade7069c15185c",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"0.5.3"
-    },
-    "patsy-tests":{
-       "depends":[
-          "patsy"
-       ],
-       "file_name":"patsy-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"patsy-tests",
-       "package_type":"package",
-       "sha256":"54a38db7a1db7e64c496ae893e5bc19a43db5b838e4e0fc8814ab402e3609ead",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"0.5.3"
-    },
-    "peewee":{
-       "depends":[
-          "sqlite3",
-          "cffi"
-       ],
-       "file_name":"peewee-3.16.3-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "peewee"
-       ],
-       "install_dir":"site",
-       "name":"peewee",
-       "package_type":"package",
-       "sha256":"eff9c450ddd23395f490437d253a880d51bda78dadcc866eb9536821eef243c4",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"3.16.3"
-    },
-    "peewee-tests":{
-       "depends":[
-          "peewee"
-       ],
-       "file_name":"peewee-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"peewee-tests",
-       "package_type":"package",
-       "sha256":"919abb68dcddccab72ffddcb4cbb23e25058b0db78c2adbeca25ef86db9e0f4d",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"3.16.3"
-    },
-    "pillow":{
-       "depends":[
-
-       ],
-       "file_name":"Pillow-10.0.0-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "PIL"
-       ],
-       "install_dir":"site",
-       "name":"Pillow",
-       "package_type":"package",
-       "sha256":"ec41e3c984df46dc6ea5153a2e62dbc9a1b67036a288ae6ddd2cb931cca7df39",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"10.0.0"
-    },
-    "pillow_heif":{
-       "depends":[
-          "cffi",
-          "pillow",
-          "libheif"
-       ],
-       "file_name":"pillow_heif-0.8.0-cp36-abi3-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "pillow_heif"
-       ],
-       "install_dir":"site",
-       "name":"pillow_heif",
-       "package_type":"package",
-       "sha256":"e7790b122e60aba187cfd1d3ec910e0235063b95b7bfbde284ac2d206a58c8e7",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"0.8.0"
-    },
-    "pkgconfig":{
-       "depends":[
-
-       ],
-       "file_name":"pkgconfig-1.5.5-py3-none-any.whl",
-       "imports":[
-          "pkgconfig"
-       ],
-       "install_dir":"site",
-       "name":"pkgconfig",
-       "package_type":"package",
-       "sha256":"eb68dbf89ce1b4ad9f1778d3f415ea7c3fdde38689a10838e8e1e4aaaf7b7401",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.5.5"
-    },
-    "pluggy":{
-       "depends":[
-
-       ],
-       "file_name":"pluggy-1.2.0-py3-none-any.whl",
-       "imports":[
-          "pluggy"
-       ],
-       "install_dir":"site",
-       "name":"pluggy",
-       "package_type":"package",
-       "sha256":"2a535450361eced5956838339504dd82baa6894a51c216c88dff4bbe8be672a0",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.2.0"
-    },
-    "protobuf":{
-       "depends":[
-
-       ],
-       "file_name":"protobuf-4.23.1-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "google"
-       ],
-       "install_dir":"site",
-       "name":"protobuf",
-       "package_type":"package",
-       "sha256":"26566ac0b42e4edd078ff866228703b8d9ddf1cea6a3e9559c232f81d9e1dd91",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"4.23.1"
-    },
-    "py":{
-       "depends":[
-
-       ],
-       "file_name":"py-1.11.0-py2.py3-none-any.whl",
-       "imports":[
-          "py"
-       ],
-       "install_dir":"site",
-       "name":"py",
-       "package_type":"package",
-       "sha256":"20b4c33073cf90b767d93335556e34895171dd16f65124c8a5219400b7128436",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.11.0"
-    },
-    "pyclipper":{
-       "depends":[
-
-       ],
-       "file_name":"pyclipper-1.3.0.post4-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "pyclipper"
-       ],
-       "install_dir":"site",
-       "name":"pyclipper",
-       "package_type":"package",
-       "sha256":"8461612adcdc611d1bd7b6cfc86d4a5e40ef11bfcd9403e857ec4740d50d6496",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.3.0.post4"
-    },
-    "pycparser":{
-       "depends":[
-
-       ],
-       "file_name":"pycparser-2.21-py2.py3-none-any.whl",
-       "imports":[
-          "pycparser"
-       ],
-       "install_dir":"site",
-       "name":"pycparser",
-       "package_type":"package",
-       "sha256":"1bf16284f2b19a6fa91d09f028a6e4a76e42f94ee2940456a1cb8fed8a1d7314",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"2.21"
-    },
-    "pycryptodome":{
-       "depends":[
-
-       ],
-       "file_name":"pycryptodome-3.18.0-cp35-abi3-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "Crypto"
-       ],
-       "install_dir":"site",
-       "name":"pycryptodome",
-       "package_type":"package",
-       "sha256":"5ae3073b2f6cee67e4f8d431e175abeb90235feeea1a316750229c652d103083",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"3.18.0"
-    },
-    "pycryptodome-tests":{
-       "depends":[
-          "pycryptodome"
-       ],
-       "file_name":"pycryptodome-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"pycryptodome-tests",
-       "package_type":"package",
-       "sha256":"7f78aed02fd5103a9cfc001d46f659351f3e1c4a5fedee9cace5d334825f0bd2",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"3.18.0"
-    },
-    "pydantic":{
-       "depends":[
-          "typing-extensions"
-       ],
-       "file_name":"pydantic-1.10.7-py3-none-any.whl",
-       "imports":[
-          "pydantic"
-       ],
-       "install_dir":"site",
-       "name":"pydantic",
-       "package_type":"package",
-       "sha256":"e8ec7048d6a09d348f153ed8bfe580a89756e930dab5310d7311ee4925172abc",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.10.7"
-    },
-    "pydecimal":{
-       "depends":[
-
-       ],
-       "file_name":"pydecimal-1.0.0.zip",
-       "imports":[
-          "_pydecimal"
-       ],
-       "install_dir":"stdlib",
-       "name":"pydecimal",
-       "package_type":"cpython_module",
-       "sha256":"9906dc6bce0a1f8fce2fb3c6b87a7a91ca6e48a7e07f4183ded4bcf5b0b446b3",
-       "shared_library":true,
-       "unvendored_tests":false,
-       "version":"1.0.0"
-    },
-    "pydoc_data":{
-       "depends":[
-
-       ],
-       "file_name":"pydoc_data-1.0.0.zip",
-       "imports":[
-          "pydoc_data"
-       ],
-       "install_dir":"stdlib",
-       "name":"pydoc_data",
-       "package_type":"cpython_module",
-       "sha256":"dba6e2ba40cf34f09e4619e04f06d9e299234c22505efa95fb4ee92a924843f0",
-       "shared_library":true,
-       "unvendored_tests":false,
-       "version":"1.0.0"
-    },
-    "pyerfa":{
-       "depends":[
-          "numpy"
-       ],
-       "file_name":"pyerfa-2.0.0.3-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "erfa"
-       ],
-       "install_dir":"site",
-       "name":"pyerfa",
-       "package_type":"package",
-       "sha256":"108281ab32585d15f113d18df016786d7e97311a0a15e12a00f886a954ab2bb2",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"2.0.0.3"
-    },
-    "pyerfa-tests":{
-       "depends":[
-          "pyerfa"
-       ],
-       "file_name":"pyerfa-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"pyerfa-tests",
-       "package_type":"package",
-       "sha256":"13a1fdd3e573a8b7f187eed54f453f767bf4e5288251efa451e234c363686f6a",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"2.0.0.3"
-    },
-    "pygments":{
-       "depends":[
-
-       ],
-       "file_name":"Pygments-2.16.1-py3-none-any.whl",
-       "imports":[
-          "pygments"
-       ],
-       "install_dir":"site",
-       "name":"Pygments",
-       "package_type":"package",
-       "sha256":"fd639d354f131822297f32468c1d0b4d4631373b893157c791723c2859714bfb",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"2.16.1"
-    },
-    "pyheif":{
-       "depends":[
-          "cffi"
-       ],
-       "file_name":"pyheif-0.7.1-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "pyheif"
-       ],
-       "install_dir":"site",
-       "name":"pyheif",
-       "package_type":"package",
-       "sha256":"155c3e1b522a43630c805b8758b991e78dcf488450f84d77e99f7a55c7aaf9fa",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"0.7.1"
-    },
-    "pyinstrument":{
-       "depends":[
-
-       ],
-       "file_name":"pyinstrument-4.4.0-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "pyinstrument"
-       ],
-       "install_dir":"site",
-       "name":"pyinstrument",
-       "package_type":"package",
-       "sha256":"378ff23a15f53f0cb0d089022bbad0fe49fe7ea971fa09ed57a535477ab76e62",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"4.4.0"
-    },
-    "pynacl":{
-       "depends":[
-          "cffi"
-       ],
-       "file_name":"PyNaCl-1.5.0-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "nacl"
-       ],
-       "install_dir":"site",
-       "name":"pynacl",
-       "package_type":"package",
-       "sha256":"5472da2430053af787c2b7b80219e6ed66b29f2d49483506feee91ebde6d0f89",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.5.0"
-    },
-    "pyodide-http":{
-       "depends":[
-
-       ],
-       "file_name":"pyodide_http-0.2.1-py3-none-any.whl",
-       "imports":[
-          "pyodide_http"
-       ],
-       "install_dir":"site",
-       "name":"pyodide-http",
-       "package_type":"package",
-       "sha256":"658afcd5b9b988bc088680a9c57d1bb6662b6fb60b40f469aaac0029d0189bc2",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"0.2.1"
-    },
-    "pyparsing":{
-       "depends":[
-
-       ],
-       "file_name":"pyparsing-3.1.1-py3-none-any.whl",
-       "imports":[
-          "pyparsing"
-       ],
-       "install_dir":"site",
-       "name":"pyparsing",
-       "package_type":"package",
-       "sha256":"66f134552b8a660c0c6d4123189de503f4b8931a1226c25fb5eee4044ce0ef7e",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"3.1.1"
-    },
-    "pyproj":{
-       "depends":[
-          "certifi",
-          "sqlite3"
-       ],
-       "file_name":"pyproj-3.4.1-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "pyproj"
-       ],
-       "install_dir":"site",
-       "name":"pyproj",
-       "package_type":"package",
-       "sha256":"703953cd99875a3a52d38acc7f92501cd18293d9cf01905004a881af621f4a96",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"3.4.1"
-    },
-    "pyrsistent":{
-       "depends":[
-
-       ],
-       "file_name":"pyrsistent-0.19.3-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "_pyrsistent_version",
-          "pyrsistent"
-       ],
-       "install_dir":"site",
-       "name":"pyrsistent",
-       "package_type":"package",
-       "sha256":"04555daa7915bed393f9c32704ae6638b507101e0952135004a5f663f7f9ca69",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"0.19.3"
-    },
-    "pysam":{
-       "depends":[
-
-       ],
-       "file_name":"pysam-0.22.0-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "pysam"
-       ],
-       "install_dir":"site",
-       "name":"pysam",
-       "package_type":"package",
-       "sha256":"0a57b514bb77df7edd688914cad4ecae932e3304a9c78f80df80fa153389857f",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"0.22.0"
-    },
-    "pyshp":{
-       "depends":[
-
-       ],
-       "file_name":"pyshp-2.3.1-py2.py3-none-any.whl",
-       "imports":[
-          "shapefile"
-       ],
-       "install_dir":"site",
-       "name":"pyshp",
-       "package_type":"package",
-       "sha256":"28ce478caf57b2863a17e8ca1dd67903babb3b2ef202ce15c32c2cc64253ea96",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"2.3.1"
-    },
-    "pytest":{
-       "depends":[
-          "atomicwrites",
-          "attrs",
-          "more-itertools",
-          "pluggy",
-          "py",
-          "setuptools",
-          "six",
-          "iniconfig",
-          "exceptiongroup"
-       ],
-       "file_name":"pytest-7.2.2-py3-none-any.whl",
-       "imports":[
-          "_pytest",
-          "pytest"
-       ],
-       "install_dir":"site",
-       "name":"pytest",
-       "package_type":"package",
-       "sha256":"10aa700d1a351061b759ea1c631e9efe685e15bdcc884331df89769d1053672b",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"7.2.2"
-    },
-    "pytest-benchmark":{
-       "depends":[
-
-       ],
-       "file_name":"pytest_benchmark-4.0.0-py3-none-any.whl",
-       "imports":[
-          "pytest_benchmark"
-       ],
-       "install_dir":"site",
-       "name":"pytest-benchmark",
-       "package_type":"package",
-       "sha256":"3b61f286d0675971a67debb48f4f3d5b8cc6757bf207b9ea639071af36b74ceb",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"4.0.0"
-    },
-    "python-dateutil":{
-       "depends":[
-          "six"
-       ],
-       "file_name":"python_dateutil-2.8.2-py2.py3-none-any.whl",
-       "imports":[
-          "dateutil"
-       ],
-       "install_dir":"site",
-       "name":"python-dateutil",
-       "package_type":"package",
-       "sha256":"0bf56aed7b371f3b5dd0e42f443cdc409a27f357bb6da344d17977e3014c3860",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"2.8.2"
-    },
-    "python-magic":{
-       "depends":[
-          "libmagic"
-       ],
-       "file_name":"python_magic-0.4.27-py2.py3-none-any.whl",
-       "imports":[
-          "magic"
-       ],
-       "install_dir":"site",
-       "name":"python-magic",
-       "package_type":"package",
-       "sha256":"e2dc33aad018a28df14a9385613b00da4eab8f2b1a022ac5b850cb28e1648e69",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"0.4.27"
-    },
-    "python-sat":{
-       "depends":[
-          "six"
-       ],
-       "file_name":"python_sat-0.1.7.dev26-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "pysat"
-       ],
-       "install_dir":"site",
-       "name":"python-sat",
-       "package_type":"package",
-       "sha256":"5c033d78c5bdf488dd22a93e563d9662cebd7a067f51b6c2f1c43b8f580e312e",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"0.1.7.dev26"
-    },
-    "python_solvespace":{
-       "depends":[
-
-       ],
-       "file_name":"python_solvespace-3.0.8-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "python_solvespace"
-       ],
-       "install_dir":"site",
-       "name":"python_solvespace",
-       "package_type":"package",
-       "sha256":"6c28bd075b97ff40caf3960ce596a5a0ffc7af2d6713cd1bfd49a264acf2d40b",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"3.0.8"
-    },
-    "pytz":{
-       "depends":[
-
-       ],
-       "file_name":"pytz-2023.3-py2.py3-none-any.whl",
-       "imports":[
-          "pytz"
-       ],
-       "install_dir":"site",
-       "name":"pytz",
-       "package_type":"package",
-       "sha256":"ff1c4fc865d6842908b2e1f46021cc6bc935f8c4484cab6f0fdaf8e9977a25de",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"2023.3"
-    },
-    "pywavelets":{
-       "depends":[
-          "distutils",
-          "numpy",
-          "matplotlib",
-          "scipy"
-       ],
-       "file_name":"PyWavelets-1.4.1-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "pywt"
-       ],
-       "install_dir":"site",
-       "name":"pywavelets",
-       "package_type":"package",
-       "sha256":"a7b2d449f15a5d6c4c9bbe2ed9a441902ed5165ead1e67236d9fdc2a409c384d",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"1.4.1"
-    },
-    "pywavelets-tests":{
-       "depends":[
-          "pywavelets"
-       ],
-       "file_name":"pywavelets-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"pywavelets-tests",
-       "package_type":"package",
-       "sha256":"6770d4953324a538e9bc320fdd1013bc045d4110dda2fffbdbd7427f38adf89f",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.4.1"
-    },
-    "pyxel":{
-       "depends":[
-
-       ],
-       "file_name":"pyxel-1.9.10-cp37-abi3-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "pyxel"
-       ],
-       "install_dir":"site",
-       "name":"pyxel",
-       "package_type":"package",
-       "sha256":"67c2849eb3bb587546a608aebde03817cadd936d8b3696b00693e3967e9a07fd",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.9.10"
-    },
-    "pyyaml":{
-       "depends":[
-
-       ],
-       "file_name":"PyYAML-6.0.1-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "_yaml",
-          "yaml"
-       ],
-       "install_dir":"site",
-       "name":"pyyaml",
-       "package_type":"package",
-       "sha256":"8157ac93d290081a2544277360443a76fc6fb6571f6946537a093a8325e41f63",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"6.0.1"
-    },
-    "rebound":{
-       "depends":[
-          "numpy"
-       ],
-       "file_name":"rebound-3.24.2-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "rebound"
-       ],
-       "install_dir":"site",
-       "name":"rebound",
-       "package_type":"package",
-       "sha256":"e5450c5094e077ed4eb1e9f0623b79d7d7ced01e0ebe0927225e9b4051f4cb18",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"3.24.2"
-    },
-    "reboundx":{
-       "depends":[
-          "rebound",
-          "numpy"
-       ],
-       "file_name":"reboundx-3.10.1-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "reboundx"
-       ],
-       "install_dir":"site",
-       "name":"reboundx",
-       "package_type":"package",
-       "sha256":"751f2eb356d3a7a5c34452abe24fceeca6306a9f6a0ac178aca4a9103dadb976",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"3.10.1"
-    },
-    "regex":{
-       "depends":[
-
-       ],
-       "file_name":"regex-2023.8.8-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "regex"
-       ],
-       "install_dir":"site",
-       "name":"regex",
-       "package_type":"package",
-       "sha256":"41c11f8c57c3f9725898ad61d5b05f06d39df14f6b512eeb485a92168823b836",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"2023.8.8"
-    },
-    "regex-tests":{
-       "depends":[
-          "regex"
-       ],
-       "file_name":"regex-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"regex-tests",
-       "package_type":"package",
-       "sha256":"06a4e804dc1b46024db4cb931de5208d6e2d1ccd8aa5e0cdc8516e947313feaf",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"2023.8.8"
-    },
-    "retrying":{
-       "depends":[
-          "six"
-       ],
-       "file_name":"retrying-1.3.4-py3-none-any.whl",
-       "imports":[
-          "retrying"
-       ],
-       "install_dir":"site",
-       "name":"retrying",
-       "package_type":"package",
-       "sha256":"fa9c2a971c150c071ff8b52f2c2ce63a7dd741ca5ca875102f1030a6cdfad544",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.3.4"
-    },
-    "river":{
-       "depends":[
-          "numpy",
-          "pandas",
-          "pytest",
-          "scipy"
-       ],
-       "file_name":"river-0.19.0-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "river"
-       ],
-       "install_dir":"site",
-       "name":"river",
-       "package_type":"package",
-       "sha256":"feafa8ecbf1208d1748de289a10ebb452d12d324c8a60548fe20131ce5cad32d",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"0.19.0"
-    },
-    "river-tests":{
-       "depends":[
-          "river"
-       ],
-       "file_name":"river-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"river-tests",
-       "package_type":"package",
-       "sha256":"9f8a1c73c12a24758d5cf274e4a0974df25b4dc2c7868ccaba3867a8589ee640",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"0.19.0"
-    },
-    "robotraconteur":{
-       "depends":[
-          "numpy"
-       ],
-       "file_name":"RobotRaconteur-0.15.1-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "RobotRaconteur"
-       ],
-       "install_dir":"site",
-       "name":"RobotRaconteur",
-       "package_type":"package",
-       "sha256":"bbc3dc43bccb1b5c6ab32c51e83c1d451894f03e80c752ca9bd5c4c36e09e622",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"0.15.1"
-    },
-    "ruamel.yaml":{
-       "depends":[
-
-       ],
-       "file_name":"ruamel.yaml-0.17.32-py3-none-any.whl",
-       "imports":[
-          "ruamel"
-       ],
-       "install_dir":"site",
-       "name":"ruamel.yaml",
-       "package_type":"package",
-       "sha256":"b19df2abe4bc0872417e21bc0dc44b9c56da991440b8f737bc502bf9408fa8f7",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"0.17.32"
-    },
-    "rust-panic-test":{
-       "depends":[
-
-       ],
-       "file_name":"rust_panic_test-1.0-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "rust-panic-test"
-       ],
-       "install_dir":"site",
-       "name":"rust-panic-test",
-       "package_type":"package",
-       "sha256":"733325f5db013c40c38ac91cc93286b014f3d142bedf430991f34f6fce9563d2",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.0"
-    },
-    "scikit-image":{
-       "depends":[
-          "packaging",
-          "numpy",
-          "scipy",
-          "networkx",
-          "pillow",
-          "imageio",
-          "pywavelets",
-          "lazy_loader"
-       ],
-       "file_name":"scikit_image-0.21.0-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "skimage"
-       ],
-       "install_dir":"site",
-       "name":"scikit-image",
-       "package_type":"package",
-       "sha256":"c1ce4d916365517a8cd2a22e551b0f390b310b71719ac5ee0afa7bd5c1e0b401",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"0.21.0"
-    },
-    "scikit-image-tests":{
-       "depends":[
-          "scikit-image"
-       ],
-       "file_name":"scikit-image-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"scikit-image-tests",
-       "package_type":"package",
-       "sha256":"e7533c586c5db492fd68aeeb82af92a2c16b532debe5cb35bf7879132d74b5a2",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"0.21.0"
-    },
-    "scikit-learn":{
-       "depends":[
-          "scipy",
-          "joblib",
-          "threadpoolctl"
-       ],
-       "file_name":"scikit_learn-1.3.1-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "sklearn"
-       ],
-       "install_dir":"site",
-       "name":"scikit-learn",
-       "package_type":"package",
-       "sha256":"e63d82e4e04234f6eeb49f0904654acd6b5764b63567f64c0500e700af98b0b0",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"1.3.1"
-    },
-    "scikit-learn-tests":{
-       "depends":[
-          "scikit-learn"
-       ],
-       "file_name":"scikit-learn-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"scikit-learn-tests",
-       "package_type":"package",
-       "sha256":"6a120cd94260ab1e93e63905bb5aa6aaeb028b8a20a1648eaac05938ee135449",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.3.1"
-    },
-    "scipy":{
-       "depends":[
-          "numpy",
-          "openblas"
-       ],
-       "file_name":"scipy-1.11.2-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "scipy"
-       ],
-       "install_dir":"site",
-       "name":"scipy",
-       "package_type":"package",
-       "sha256":"94e5c43b9e9e877b9ea991102ba0a6f9b4aa5b8ebbefcc27a66e50226cb6aa43",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"1.11.2"
-    },
-    "scipy-tests":{
-       "depends":[
-          "scipy"
-       ],
-       "file_name":"scipy-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"scipy-tests",
-       "package_type":"package",
-       "sha256":"96872c6fc60aff93da8c1b40197b2de54c0d3a0b24021c4493a3a023bbd16274",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.11.2"
-    },
-    "screed":{
-       "depends":[
-
-       ],
-       "file_name":"screed-1.1.2-py2.py3-none-any.whl",
-       "imports":[
-          "bigtests",
-          "screed"
-       ],
-       "install_dir":"site",
-       "name":"screed",
-       "package_type":"package",
-       "sha256":"21d86f53256e5d51ba44a9fd1d64a63994d0f7ecf954d8413c343846d97b304d",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"1.1.2"
-    },
-    "screed-tests":{
-       "depends":[
-          "screed"
-       ],
-       "file_name":"screed-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"screed-tests",
-       "package_type":"package",
-       "sha256":"90a0ff98ee61ceba88ca3570fae3eb078c6b480ee2759d20d1cfafb2ce7381ca",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.1.2"
-    },
-    "setuptools":{
-       "depends":[
-          "distutils",
-          "pyparsing"
-       ],
-       "file_name":"setuptools-68.1.2-py3-none-any.whl",
-       "imports":[
-          "_distutils_hack",
-          "pkg_resources",
-          "setuptools"
-       ],
-       "install_dir":"site",
-       "name":"setuptools",
-       "package_type":"package",
-       "sha256":"bc6795827a6332eea8b78943ff939f0e672cab2a779c8364255665f58b54fc95",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"68.1.2"
-    },
-    "shapely":{
-       "depends":[
-          "numpy"
-       ],
-       "file_name":"Shapely-1.8.2-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "shapely"
-       ],
-       "install_dir":"site",
-       "name":"shapely",
-       "package_type":"package",
-       "sha256":"e9743338fa38d23f9ef58251c5e6b82359c3dd62705107d686458dbbe968e30d",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"1.8.2"
-    },
-    "shapely-tests":{
-       "depends":[
-          "shapely"
-       ],
-       "file_name":"shapely-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"shapely-tests",
-       "package_type":"package",
-       "sha256":"53cbb821363dec1009d297e64ae628db81fb45b193eebbf9ec8ece0e0e619e86",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.8.2"
-    },
-    "sharedlib-test":{
-       "depends":[
-
-       ],
-       "file_name":"sharedlib-test-1.0.zip",
-       "imports":[
-
-       ],
-       "install_dir":"dynlib",
-       "name":"sharedlib-test",
-       "package_type":"shared_library",
-       "sha256":"dcd78a95eb0c4f4b1f638e5ee1565857b5ea750d4aef076fef963de578b356d4",
-       "shared_library":true,
-       "unvendored_tests":false,
-       "version":"1.0"
-    },
-    "sharedlib-test-py":{
-       "depends":[
-          "sharedlib-test"
-       ],
-       "file_name":"sharedlib_test_py-1.0-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "sharedlib_test"
-       ],
-       "install_dir":"site",
-       "name":"sharedlib-test-py",
-       "package_type":"package",
-       "sha256":"28eb32fd58336f8844a875b87d42c20edaf9d2f120708f7581ed37574f5d941a",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.0"
-    },
-    "simplejson":{
-       "depends":[
-
-       ],
-       "file_name":"simplejson-3.19.1-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "simplejson"
-       ],
-       "install_dir":"site",
-       "name":"simplejson",
-       "package_type":"package",
-       "sha256":"c0dfd2948e07743c957196449fe04903f12e0a49cd3dc6cdae9a936d4ddb8289",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"3.19.1"
-    },
-    "simplejson-tests":{
-       "depends":[
-          "simplejson"
-       ],
-       "file_name":"simplejson-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"simplejson-tests",
-       "package_type":"package",
-       "sha256":"58ce47e899fe3d52ad4166d9e13853dac28d068c847788085d32dc3286c7a0af",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"3.19.1"
-    },
-    "sisl":{
-       "depends":[
-          "pyparsing",
-          "numpy",
-          "scipy",
-          "tqdm",
-          "xarray",
-          "pandas",
-          "matplotlib"
-       ],
-       "file_name":"sisl-0.14.2-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "sisl_toolbox",
-          "sisl"
-       ],
-       "install_dir":"site",
-       "name":"sisl",
-       "package_type":"package",
-       "sha256":"af2da3845b64b9e3509fe159cfa285f40dc3d4a0620e72abbf356c2feb6e37d6",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"0.14.2"
-    },
-    "sisl-tests":{
-       "depends":[
-          "sisl"
-       ],
-       "file_name":"sisl-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"sisl-tests",
-       "package_type":"package",
-       "sha256":"3053f108f5f5c010f7d3e2e98f8e8f3e5010a5eef66365419b799af8eb84707b",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"0.14.2"
-    },
-    "six":{
-       "depends":[
-
-       ],
-       "file_name":"six-1.16.0-py2.py3-none-any.whl",
-       "imports":[
-          "six"
-       ],
-       "install_dir":"site",
-       "name":"six",
-       "package_type":"package",
-       "sha256":"af66f234ee7a11352e1f5e7f28d134b9aef03183c8b7e57838c8fd0670ffb69c",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.16.0"
-    },
-    "smart_open":{
-       "depends":[
-
-       ],
-       "file_name":"smart_open-6.3.0-py3-none-any.whl",
-       "imports":[
-          "smart_open"
-       ],
-       "install_dir":"site",
-       "name":"smart_open",
-       "package_type":"package",
-       "sha256":"45783a97bbadc6c574d5f20c69979327d5d7e5a781e49ffd702a409fa83ff2d8",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"6.3.0"
-    },
-    "soupsieve":{
-       "depends":[
-
-       ],
-       "file_name":"soupsieve-2.4.1-py3-none-any.whl",
-       "imports":[
-          "soupsieve"
-       ],
-       "install_dir":"site",
-       "name":"soupsieve",
-       "package_type":"package",
-       "sha256":"975d127ea2dd9c024b607e08a0dbaf06bd97c618668f8558a5d6f0abf733a18f",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"2.4.1"
-    },
-    "sourmash":{
-       "depends":[
-          "screed",
-          "cffi",
-          "deprecation",
-          "cachetools",
-          "numpy",
-          "matplotlib",
-          "scipy",
-          "sqlite3",
-          "bitstring"
-       ],
-       "file_name":"sourmash-4.8.4-py3-none-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "sourmash"
-       ],
-       "install_dir":"site",
-       "name":"sourmash",
-       "package_type":"package",
-       "sha256":"92b42cb1e99fbe8a0a89cd0fea803ad9bbe2d5a598cf5abfe6f8c72c19594ee3",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"4.8.4"
-    },
-    "sparseqr":{
-       "depends":[
-          "pycparser",
-          "cffi",
-          "numpy",
-          "scipy",
-          "suitesparse",
-          "distutils"
-       ],
-       "file_name":"sparseqr-1.2-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "sparseqr"
-       ],
-       "install_dir":"site",
-       "name":"sparseqr",
-       "package_type":"package",
-       "sha256":"f8b804383ef963cba4d891b2e5f1475e96dc187c629e9c07024ce5b5c9ed356a",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.2"
-    },
-    "sqlalchemy":{
-       "depends":[
-          "sqlite3",
-          "typing-extensions"
-       ],
-       "file_name":"SQLAlchemy-2.0.20-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "sqlalchemy"
-       ],
-       "install_dir":"site",
-       "name":"sqlalchemy",
-       "package_type":"package",
-       "sha256":"26d916aec88466f14e872d48541cd61adf52b5a0e2f04c4e689036aa6cb27dbf",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"2.0.20"
-    },
-    "sqlalchemy-tests":{
-       "depends":[
-          "sqlalchemy"
-       ],
-       "file_name":"sqlalchemy-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"sqlalchemy-tests",
-       "package_type":"package",
-       "sha256":"d20dacde41cb32f60bebaa883936e7417add9db2f2ef26bd877cb0182942d153",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"2.0.20"
-    },
-    "sqlite3":{
-       "depends":[
-
-       ],
-       "file_name":"sqlite3-1.0.0.zip",
-       "imports":[
-          "sqlite3",
-          "_sqlite3"
-       ],
-       "install_dir":"stdlib",
-       "name":"sqlite3",
-       "package_type":"cpython_module",
-       "sha256":"af3dcc095eaf30d9d0b6ac318b780ec53e5ba73d7b86d3b56c9585a08c25fc1c",
-       "shared_library":true,
-       "unvendored_tests":false,
-       "version":"1.0.0"
-    },
-    "ssl":{
-       "depends":[
-          "openssl"
-       ],
-       "file_name":"ssl-1.0.0.zip",
-       "imports":[
-          "ssl",
-          "_ssl"
-       ],
-       "install_dir":"stdlib",
-       "name":"ssl",
-       "package_type":"cpython_module",
-       "sha256":"9088d85ceb960416cce39aab4c2144c0934efa2175bf50573c24aed93db360c9",
-       "shared_library":true,
-       "unvendored_tests":false,
-       "version":"1.0.0"
-    },
-    "statsmodels":{
-       "depends":[
-          "distutils",
-          "numpy",
-          "scipy",
-          "pandas",
-          "patsy",
-          "packaging"
-       ],
-       "file_name":"statsmodels-0.14.0-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "statsmodels"
-       ],
-       "install_dir":"site",
-       "name":"statsmodels",
-       "package_type":"package",
-       "sha256":"09aeeef9344333e6868846aeaae64be7c88bfb919fb3d9de98a320bf0e6d64f6",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"0.14.0"
-    },
-    "statsmodels-tests":{
-       "depends":[
-          "statsmodels"
-       ],
-       "file_name":"statsmodels-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"statsmodels-tests",
-       "package_type":"package",
-       "sha256":"c6dd610d90324a101517aa69c5ad32f1f4c0c371260a28bd485d02def3176e72",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"0.14.0"
-    },
-    "suitesparse":{
-       "depends":[
-          "openblas"
-       ],
-       "file_name":"suitesparse-5.11.0.zip",
-       "imports":[
-
-       ],
-       "install_dir":"dynlib",
-       "name":"suitesparse",
-       "package_type":"shared_library",
-       "sha256":"5dbea0b8df376de1c6732ac1f08338022be6fe119028411ac66595ca575f6769",
-       "shared_library":true,
-       "unvendored_tests":false,
-       "version":"5.11.0"
-    },
-    "svgwrite":{
-       "depends":[
-
-       ],
-       "file_name":"svgwrite-1.4.3-py3-none-any.whl",
-       "imports":[
-          "svgwrite"
-       ],
-       "install_dir":"site",
-       "name":"svgwrite",
-       "package_type":"package",
-       "sha256":"95b206dc1a5977b054559fc95d7b32eb7385036b79c64ba95f0c8c5045e86fd5",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.4.3"
-    },
-    "swiglpk":{
-       "depends":[
-
-       ],
-       "file_name":"swiglpk-5.0.8-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "swiglpk"
-       ],
-       "install_dir":"site",
-       "name":"swiglpk",
-       "package_type":"package",
-       "sha256":"3a3a342a2723faa155da7da680506b911f8750002e9245fe017d43f788b14d88",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"5.0.8"
-    },
-    "sympy":{
-       "depends":[
-          "distutils",
-          "mpmath"
-       ],
-       "file_name":"sympy-1.12-py3-none-any.whl",
-       "imports":[
-          "isympy",
-          "sympy"
-       ],
-       "install_dir":"site",
-       "name":"sympy",
-       "package_type":"package",
-       "sha256":"ef0d3e829d41757b6de728380f9b0684151ce2c2070498c635b32a4177a2bc83",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"1.12"
-    },
-    "sympy-tests":{
-       "depends":[
-          "sympy"
-       ],
-       "file_name":"sympy-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"sympy-tests",
-       "package_type":"package",
-       "sha256":"d95ca3cb38ef68ea05a7e669035d2bf51eab417357b27f4b3943c1c3a5d6592e",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.12"
-    },
-    "tblib":{
-       "depends":[
-
-       ],
-       "file_name":"tblib-3.0.0-py3-none-any.whl",
-       "imports":[
-          "tblib"
-       ],
-       "install_dir":"site",
-       "name":"tblib",
-       "package_type":"package",
-       "sha256":"9a7ff2c2aba05a32411b5381417c5f5f1f68dddd195b5f27241d3bc6be4c72c7",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"3.0.0"
-    },
-    "termcolor":{
-       "depends":[
-
-       ],
-       "file_name":"termcolor-2.3.0-py3-none-any.whl",
-       "imports":[
-          "termcolor"
-       ],
-       "install_dir":"site",
-       "name":"termcolor",
-       "package_type":"package",
-       "sha256":"84534b40c4bd1d5d576412f2bb3c8c942beddafab11fe03363d7074be5886f7d",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"2.3.0"
-    },
-    "test":{
-       "depends":[
-
-       ],
-       "file_name":"test-1.0.0.zip",
-       "imports":[
-          "test"
-       ],
-       "install_dir":"stdlib",
-       "name":"test",
-       "package_type":"cpython_module",
-       "sha256":"770290c458238159d25ce7cf55c3c22961297e20c8c260b4fd490682db3eeda4",
-       "shared_library":true,
-       "unvendored_tests":false,
-       "version":"1.0.0"
-    },
-    "texttable":{
-       "depends":[
-
-       ],
-       "file_name":"texttable-1.6.7-py2.py3-none-any.whl",
-       "imports":[
-          "texttable"
-       ],
-       "install_dir":"site",
-       "name":"texttable",
-       "package_type":"package",
-       "sha256":"f6a7bb37320f4a1fc4c0f315609d2d071d497f57c57932bc2ff79da9c1c9abc4",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.6.7"
-    },
-    "threadpoolctl":{
-       "depends":[
-
-       ],
-       "file_name":"threadpoolctl-3.2.0-py3-none-any.whl",
-       "imports":[
-          "threadpoolctl"
-       ],
-       "install_dir":"site",
-       "name":"threadpoolctl",
-       "package_type":"package",
-       "sha256":"b5e0af61311a4e4c1fe875dc880e9f668817c0eea15a0a35ed1c81269b64aca2",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"3.2.0"
-    },
-    "tomli":{
-       "depends":[
-
-       ],
-       "file_name":"tomli-2.0.1-py3-none-any.whl",
-       "imports":[
-          "tomli"
-       ],
-       "install_dir":"site",
-       "name":"tomli",
-       "package_type":"package",
-       "sha256":"c8a1bc2f9306b0f2549afe51ec6f9e2daf8fb16cc32502316ed25c0596ec9490",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"2.0.1"
-    },
-    "tomli-w":{
-       "depends":[
-
-       ],
-       "file_name":"tomli_w-1.0.0-py3-none-any.whl",
-       "imports":[
-          "tomli_w"
-       ],
-       "install_dir":"site",
-       "name":"tomli-w",
-       "package_type":"package",
-       "sha256":"382bc8caa39b519229408c847ba9f471defef738116431c508e93b877500b051",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.0.0"
-    },
-    "toolz":{
-       "depends":[
-
-       ],
-       "file_name":"toolz-0.12.0-py3-none-any.whl",
-       "imports":[
-          "tlz",
-          "toolz"
-       ],
-       "install_dir":"site",
-       "name":"toolz",
-       "package_type":"package",
-       "sha256":"7fa1b8935a7313656499ef7eb6b327f073db5c77f7f2b19667f249b24c8f557d",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"0.12.0"
-    },
-    "toolz-tests":{
-       "depends":[
-          "toolz"
-       ],
-       "file_name":"toolz-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"toolz-tests",
-       "package_type":"package",
-       "sha256":"f67b2147cad877d7d787177fc7ff74d919ab7daa2fbe85426c9ff2f4557fa0fe",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"0.12.0"
-    },
-    "tqdm":{
-       "depends":[
-
-       ],
-       "file_name":"tqdm-4.66.1-py3-none-any.whl",
-       "imports":[
-          "tqdm"
-       ],
-       "install_dir":"site",
-       "name":"tqdm",
-       "package_type":"package",
-       "sha256":"ea144d101ec8ef0a8858cc7a802f4970f59593f4ac28bd96b40d7fd61e5294b1",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"4.66.1"
-    },
-    "traits":{
-       "depends":[
-
-       ],
-       "file_name":"traits-6.4.2-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "traits"
-       ],
-       "install_dir":"site",
-       "name":"traits",
-       "package_type":"package",
-       "sha256":"55c5c2a095992a101ac5b9b76fe93b9a178b744694095262e8966e39bbfe258a",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"6.4.2"
-    },
-    "traits-tests":{
-       "depends":[
-          "traits"
-       ],
-       "file_name":"traits-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"traits-tests",
-       "package_type":"package",
-       "sha256":"e405b6beb8fcf23d9c0a5debd4ec35529f9e8ae0d802daca2e1e6d70228ef41d",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"6.4.2"
-    },
-    "tskit":{
-       "depends":[
-          "numpy",
-          "svgwrite",
-          "jsonschema"
-       ],
-       "file_name":"tskit-0.5.5-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "tskit"
-       ],
-       "install_dir":"site",
-       "name":"tskit",
-       "package_type":"package",
-       "sha256":"4b2ba5946c79285dfbb32ede79f26fcd7242f5a03af969b1a34fda1edbd1db66",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"0.5.5"
-    },
-    "typing-extensions":{
-       "depends":[
-
-       ],
-       "file_name":"typing_extensions-4.7.1-py3-none-any.whl",
-       "imports":[
-          "typing_extensions"
-       ],
-       "install_dir":"site",
-       "name":"typing-extensions",
-       "package_type":"package",
-       "sha256":"f209b05fb1299486306cbe726632c3f2675201bb05c4bc2eab8ee29ee1480c8f",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"4.7.1"
-    },
-    "uncertainties":{
-       "depends":[
-          "future"
-       ],
-       "file_name":"uncertainties-3.1.7-py2.py3-none-any.whl",
-       "imports":[
-          "uncertainties"
-       ],
-       "install_dir":"site",
-       "name":"uncertainties",
-       "package_type":"package",
-       "sha256":"48dae9e23dd68ae19d18433a71b5c2cb3f94d36dc8363828d885c5ef833a8909",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"3.1.7"
-    },
-    "uncertainties-tests":{
-       "depends":[
-          "uncertainties"
-       ],
-       "file_name":"uncertainties-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"uncertainties-tests",
-       "package_type":"package",
-       "sha256":"28256453a1bc3711c633abcbae7e9427677e1941296f8b40a26e3614ce46eef9",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"3.1.7"
-    },
-    "unyt":{
-       "depends":[
-          "numpy",
-          "sympy"
-       ],
-       "file_name":"unyt-2.9.5-py2.py3-none-any.whl",
-       "imports":[
-          "unyt"
-       ],
-       "install_dir":"site",
-       "name":"unyt",
-       "package_type":"package",
-       "sha256":"6f28e07ee05ea58760275ed0bed9190526015ef83199dae4c2b79b06c023438b",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"2.9.5"
-    },
-    "unyt-tests":{
-       "depends":[
-          "unyt"
-       ],
-       "file_name":"unyt-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"unyt-tests",
-       "package_type":"package",
-       "sha256":"2f98183db96445bccab4e824a80b23022cab70c58e558ee18913f38972f64d65",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"2.9.5"
-    },
-    "webencodings":{
-       "depends":[
-
-       ],
-       "file_name":"webencodings-0.5.1-py2.py3-none-any.whl",
-       "imports":[
-          "webencodings"
-       ],
-       "install_dir":"site",
-       "name":"webencodings",
-       "package_type":"package",
-       "sha256":"6701acc9bfc9e931ccdf1eb0831ca4167122c05ac2ffdb2ef4b2f44862cd4e70",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"0.5.1"
-    },
-    "wordcloud":{
-       "depends":[
-          "matplotlib"
-       ],
-       "file_name":"wordcloud-1.9.2-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "wordcloud"
-       ],
-       "install_dir":"site",
-       "name":"wordcloud",
-       "package_type":"package",
-       "sha256":"c9b0b2e49a79fa63ed295cb9cf850730cedd37b4155669cd6df6ae1321842415",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.9.2"
-    },
-    "wrapt":{
-       "depends":[
-
-       ],
-       "file_name":"wrapt-1.15.0-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "wrapt"
-       ],
-       "install_dir":"site",
-       "name":"wrapt",
-       "package_type":"package",
-       "sha256":"5752fb3f421b6bc77d5b2ea28b43c4104782311a12424fc9b4e94ee7f7fa15a5",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.15.0"
-    },
-    "xarray":{
-       "depends":[
-          "numpy",
-          "packaging",
-          "pandas"
-       ],
-       "file_name":"xarray-2023.12.0-py3-none-any.whl",
-       "imports":[
-          "xarray"
-       ],
-       "install_dir":"site",
-       "name":"xarray",
-       "package_type":"package",
-       "sha256":"d8c03afc94699cd417b2da23d6cd17a3cf41f6e340e77de7f3932dad4903201e",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"2023.12.0"
-    },
-    "xarray-tests":{
-       "depends":[
-          "xarray"
-       ],
-       "file_name":"xarray-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"xarray-tests",
-       "package_type":"package",
-       "sha256":"2872d95a846f07bdc867808c41b833be494c33e2f1e439012a68f6617b54e7e4",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"2023.12.0"
-    },
-    "xgboost":{
-       "depends":[
-          "numpy",
-          "scipy",
-          "setuptools"
-       ],
-       "file_name":"xgboost-1.6.1-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "xgboost"
-       ],
-       "install_dir":"site",
-       "name":"xgboost",
-       "package_type":"package",
-       "sha256":"811b3c1834407172206dd25f335839d9667ba9698e31ba9f8ea89d58b025e79c",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.6.1"
-    },
-    "xlrd":{
-       "depends":[
-
-       ],
-       "file_name":"xlrd-2.0.1-py2.py3-none-any.whl",
-       "imports":[
-          "xlrd"
-       ],
-       "install_dir":"site",
-       "name":"xlrd",
-       "package_type":"package",
-       "sha256":"101d97830986609d629998e9b817823675e4a032ebd74ba53570f1bcef26b39d",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"2.0.1"
-    },
-    "xyzservices":{
-       "depends":[
-
-       ],
-       "file_name":"xyzservices-2023.7.0-py3-none-any.whl",
-       "imports":[
-          "xyzservices"
-       ],
-       "install_dir":"site",
-       "name":"xyzservices",
-       "package_type":"package",
-       "sha256":"e8e5c22a8d243fd7a31f75771008ccf96db1beb2f4743fbaa0535fad72e65470",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"2023.7.0"
-    },
-    "xyzservices-tests":{
-       "depends":[
-          "xyzservices"
-       ],
-       "file_name":"xyzservices-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"xyzservices-tests",
-       "package_type":"package",
-       "sha256":"3f0374173170f7bd6fedbd7419f6d873472eb0a2e42cd87065a85800220be0f5",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"2023.7.0"
-    },
-    "yarl":{
-       "depends":[
-          "multidict",
-          "idna"
-       ],
-       "file_name":"yarl-1.9.2-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "yarl"
-       ],
-       "install_dir":"site",
-       "name":"yarl",
-       "package_type":"package",
-       "sha256":"f90a986be6e14dd1f93ec05bc69dec016af5f40a94e6932317640189cdca07ac",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.9.2"
-    },
-    "yt":{
-       "depends":[
-          "numpy",
-          "matplotlib",
-          "sympy",
-          "setuptools",
-          "packaging",
-          "unyt",
-          "cmyt",
-          "colorspacious",
-          "tqdm",
-          "tomli",
-          "tomli-w"
-       ],
-       "file_name":"yt-4.1.4-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "yt"
-       ],
-       "install_dir":"site",
-       "name":"yt",
-       "package_type":"package",
-       "sha256":"a02e1e23cba66e9c4ad06b5181b8bc0641e45710e95c2ab5a0d60ac2dec55675",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"4.1.4"
-    },
-    "zarr":{
-       "depends":[
-          "numpy",
-          "asciitree",
-          "numcodecs"
-       ],
-       "file_name":"zarr-2.13.3-py3-none-any.whl",
-       "imports":[
-          "zarr"
-       ],
-       "install_dir":"site",
-       "name":"zarr",
-       "package_type":"package",
-       "sha256":"4daa37039dd470cbdb4ca9bcb443ebe7bde6c1cd83b46eecca09e3bb516ef1a3",
-       "shared_library":false,
-       "unvendored_tests":true,
-       "version":"2.13.3"
-    },
-    "zarr-tests":{
-       "depends":[
-          "zarr"
-       ],
-       "file_name":"zarr-tests.tar",
-       "imports":[
-
-       ],
-       "install_dir":"site",
-       "name":"zarr-tests",
-       "package_type":"package",
-       "sha256":"9475516162c08bde9870377e3cad2c60cd9a85f5318e9d6e323e0ee1ef23d715",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"2.13.3"
-    },
-    "zengl":{
-       "depends":[
-
-       ],
-       "file_name":"zengl-1.16.0-cp311-cp311-emscripten_3_1_45_wasm32.whl",
-       "imports":[
-          "zengl",
-          "_zengl"
-       ],
-       "install_dir":"site",
-       "name":"zengl",
-       "package_type":"package",
-       "sha256":"c9ab9f7770c47c4c5cf845a924717832bb2a596baae60f4415d8bcda193bbb11",
-       "shared_library":false,
-       "unvendored_tests":false,
-       "version":"1.16.0"
-    }
- }
-};
